feat(web): migrate labs editor UI into production /editor
