+++ apps/web/src/pages/editor/index.tsx
@@ -6,43 +6,82 @@ import {
   useState,
   type FormEvent,
   type KeyboardEvent as ReactKeyboardEvent,
-} from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { DevErrorBoundary } from '../../components/DevErrorBoundary';
-import { Navbar } from '../../components/layout/Navbar';
-import { MobileBottomNav } from '../../components/layout/MobileBottomNav';
-import { Card } from '../../components/common/Card';
-import { ToastBanner } from '../../components/common/ToastBanner';
-import { useBackendUser } from '../../hooks/useBackendUser';
-import { useCreateTask, useDeleteTask, useUpdateTask, useUserTasks } from '../../hooks/useUserTasks';
-import { useDifficulties, usePillars, useStats, useTraits } from '../../hooks/useCatalogs';
-import { type UserTask } from '../../lib/api';
-import { fetchCatalogStats, fetchCatalogTraits, type Pillar } from '../../lib/api/catalogs';
-import { useAppMode } from '../../hooks/useAppMode';
-import { useDailyQuestReadiness } from '../../hooks/useDailyQuestReadiness';
-import { useOnboardingEditorNudge } from '../../hooks/useOnboardingEditorNudge';
-import { usePostLoginLanguage } from '../../i18n/postLoginLanguage';
-import { localizeDifficultyLabel, localizePillarLabel, localizeTraitLabel } from './labelLocalization';
+} from "react";
+import { Link, useLocation } from "react-router-dom";
+import { DevErrorBoundary } from "../../components/DevErrorBoundary";
+import { Navbar } from "../../components/layout/Navbar";
+import { MobileBottomNav } from "../../components/layout/MobileBottomNav";
+import { Card } from "../../components/common/Card";
+import { ToastBanner } from "../../components/common/ToastBanner";
+import { useBackendUser } from "../../hooks/useBackendUser";
+import {
+  useCreateTask,
+  useDeleteTask,
+  useUpdateTask,
+  useUserTasks,
+} from "../../hooks/useUserTasks";
+import {
+  useDifficulties,
+  usePillars,
+  useStats,
+  useTraits,
+} from "../../hooks/useCatalogs";
+import { classifyUserTask, type UserTask, type UserTaskClassification } from "../../lib/api";
+import {
+  fetchCatalogStats,
+  fetchCatalogTraits,
+  type Pillar,
+  type Trait,
+} from "../../lib/api/catalogs";
+import { useAppMode } from "../../hooks/useAppMode";
+import { useDailyQuestReadiness } from "../../hooks/useDailyQuestReadiness";
+import { useOnboardingEditorNudge } from "../../hooks/useOnboardingEditorNudge";
+import { usePostLoginLanguage } from "../../i18n/postLoginLanguage";
+import {
+  localizeDifficultyLabel,
+  localizePillarLabel,
+  localizeTraitLabel,
+} from "./labelLocalization";
+import { EDITOR_LAB_QUICK_START_SEED } from "../labs/editorLabSeed";
+import {
+  EditorGuideOverlay,
+  markEditorGuideAsSeen,
+  shouldAutoOpenEditorGuide,
+} from "../labs/editor-guide/EditorGuideOverlay";
+import type { EditorGuideStepId } from "../labs/editor-guide/guideConfig";
 import {
   getActiveSection,
   getDashboardSectionConfig,
   getDashboardSections,
   type DashboardSectionConfig,
-} from '../dashboardSections';
+} from "../dashboardSections";
 
 export const FEATURE_TASK_EDITOR_MOBILE_LIST_V1 = true;
+const ENABLE_EDITOR_GUIDE_AUTO_OPEN = true;
 
 export default function TaskEditorPage() {
   const location = useLocation();
   const { language, t } = usePostLoginLanguage();
-  const activeLocale = language === 'es' ? 'es' : 'en';
+  const activeLocale = language === "es" ? "es" : "en";
   const sections = getDashboardSections(location.pathname, language);
   const activeSection = getActiveSection(location.pathname, sections, language);
-  const taskEditorSection = getDashboardSectionConfig('editor', location.pathname, language);
-  const { backendUserId, status: backendStatus, error: backendError, reload: reloadProfile } =
-    useBackendUser();
-  const { tasks, status: tasksStatus, error: tasksError, reload: reloadTasks } =
-    useUserTasks(backendUserId);
+  const taskEditorSection = getDashboardSectionConfig(
+    "editor",
+    location.pathname,
+    language,
+  );
+  const {
+    backendUserId,
+    status: backendStatus,
+    error: backendError,
+    reload: reloadProfile,
+  } = useBackendUser();
+  const {
+    tasks,
+    status: tasksStatus,
+    error: tasksError,
+    reload: reloadTasks,
+  } = useUserTasks(backendUserId);
   const {
     data: pillars,
     isLoading: isLoadingPillars,
@@ -52,36 +91,56 @@ export default function TaskEditorPage() {
   const { data: difficulties } = useDifficulties();
   const isAppMode = useAppMode();
 
-  const [traitCatalogById, setTraitCatalogById] = useState<Record<string, { name: string; code: string }>>({});
-  const [statNamesById, setStatNamesById] = useState<Record<string, string>>({});
+  const [traitCatalogById, setTraitCatalogById] = useState<
+    Record<string, { name: string; code: string }>
+  >({});
+  const [statNamesById, setStatNamesById] = useState<Record<string, string>>(
+    {},
+  );
 
-  const [searchTerm, setSearchTerm] = useState('');
-  const [selectedPillar, setSelectedPillar] = useState('');
+  const [searchTerm, setSearchTerm] = useState("");
+  const [selectedPillar, setSelectedPillar] = useState("");
   const [showCreateModal, setShowCreateModal] = useState(false);
   const [taskToEdit, setTaskToEdit] = useState<UserTask | null>(null);
-  const [editVariant, setEditVariant] = useState<'modal' | 'panel'>('modal');
+  const [editVariant, setEditVariant] = useState<"modal" | "panel">("modal");
   const [editGroupKey, setEditGroupKey] = useState<string | null>(null);
   const [taskToDelete, setTaskToDelete] = useState<UserTask | null>(null);
-  const [deleteErrorMessage, setDeleteErrorMessage] = useState<string | null>(null);
+  const [deleteErrorMessage, setDeleteErrorMessage] = useState<string | null>(
+    null,
+  );
   const [pageToast, setPageToast] = useState<ToastMessage | null>(null);
-  const [duplicatingTaskId, setDuplicatingTaskId] = useState<string | null>(null);
+  const [duplicatingTaskId, setDuplicatingTaskId] = useState<string | null>(
+    null,
+  );
+  const [showGuideModal, setShowGuideModal] = useState(false);
+  const [activeGuideStepId, setActiveGuideStepId] =
+    useState<EditorGuideStepId>("wheel-core");
+  const [showSuggestionsModal, setShowSuggestionsModal] = useState(false);
+  const [isApplyingSuggestions, setIsApplyingSuggestions] = useState(false);
 
   const editorTopRef = useRef<HTMLDivElement | null>(null);
-  const dailyQuestReadiness = useDailyQuestReadiness(backendUserId ?? '', {
+  const dailyQuestReadiness = useDailyQuestReadiness(backendUserId ?? "", {
     enabled: Boolean(backendUserId),
   });
   const onboardingEditorNudge = useOnboardingEditorNudge({
     completedFirstDailyQuest: dailyQuestReadiness.completedFirstDailyQuest,
   });
-  const { firstEditDone, shouldShowInlineNotice, shouldShowDashboardDot, markFirstEditDone } = onboardingEditorNudge;
+  const {
+    firstEditDone,
+    shouldShowInlineNotice,
+    shouldShowDashboardDot,
+    markFirstEditDone,
+  } = onboardingEditorNudge;
 
   const { deleteTask, status: deleteStatus } = useDeleteTask();
-  const { createTask: duplicateTask } = useCreateTask();
+  const { createTask } = useCreateTask();
 
   const normalizedSearch = searchTerm.trim().toLowerCase();
-  const isBackendReady = backendStatus === 'success' && Boolean(backendUserId);
+  const isBackendReady = backendStatus === "success" && Boolean(backendUserId);
   const isLoadingTasks =
-    !isBackendReady || tasksStatus === 'loading' || (isBackendReady && tasksStatus === 'idle');
+    !isBackendReady ||
+    tasksStatus === "loading" ||
+    (isBackendReady && tasksStatus === "idle");
   const combinedError = backendError ?? tasksError;
 
   useEffect(() => {
@@ -89,6 +148,15 @@ export default function TaskEditorPage() {
       setDeleteErrorMessage(null);
     }
   }, [taskToDelete]);
+
+  useEffect(() => {
+    if (!ENABLE_EDITOR_GUIDE_AUTO_OPEN) {
+      return;
+    }
+    if (shouldAutoOpenEditorGuide()) {
+      setShowGuideModal(true);
+    }
+  }, []);
 
   useEffect(() => {
     const missingTraitsByPillar = new Map<string, Set<string>>();
@@ -122,7 +190,7 @@ export default function TaskEditorPage() {
             updates[trait.id] = { name: trait.name, code: trait.code };
           }
         } catch (error) {
-          console.error('Failed to load traits for pillar', pillarId, error);
+          console.error("Failed to load traits for pillar", pillarId, error);
         }
       }
 
@@ -170,7 +238,7 @@ export default function TaskEditorPage() {
             updates[stat.id] = stat.name;
           }
         } catch (error) {
-          console.error('Failed to load stats for trait', traitId, error);
+          console.error("Failed to load stats for trait", traitId, error);
         }
       }
 
@@ -188,7 +256,7 @@ export default function TaskEditorPage() {
 
   const pillarOptions = useMemo(() => {
     return [
-      { value: '', label: t('editor.filters.pillars.all') },
+      { value: "", label: t("editor.filters.pillars.all") },
       ...pillars.map((pillar) => ({
         value: pillar.id,
         label: localizePillarLabel(pillar.name, language),
@@ -209,16 +277,25 @@ export default function TaskEditorPage() {
     for (const [traitId, trait] of Object.entries(traitCatalogById)) {
       map.set(
         traitId,
-        localizeTraitLabel({ name: trait.name, code: trait.code, fallback: traitId }, language),
+        localizeTraitLabel(
+          { name: trait.name, code: trait.code, fallback: traitId },
+          language,
+        ),
       );
     }
     return map;
   }, [language, traitCatalogById]);
-  const statNamesMap = useMemo(() => new Map(Object.entries(statNamesById)), [statNamesById]);
+  const statNamesMap = useMemo(
+    () => new Map(Object.entries(statNamesById)),
+    [statNamesById],
+  );
   const difficultyNamesById = useMemo(() => {
     const map = new Map<string, string>();
     for (const difficulty of difficulties) {
-      map.set(difficulty.id, localizeDifficultyLabel(difficulty.name, language));
+      map.set(
+        difficulty.id,
+        localizeDifficultyLabel(difficulty.name, language),
+      );
     }
     return map;
   }, [difficulties, language]);
@@ -252,9 +329,11 @@ export default function TaskEditorPage() {
   const filteredTasks = useMemo(() => {
     return tasks.filter((task) => {
       const matchesSearch =
-        normalizedSearch.length === 0 || task.title.toLowerCase().includes(normalizedSearch);
+        normalizedSearch.length === 0 ||
+        task.title.toLowerCase().includes(normalizedSearch);
       const matchesPillar =
-        selectedPillar.length === 0 || (task.pillarId ?? '').toLowerCase() === selectedPillar.toLowerCase();
+        selectedPillar.length === 0 ||
+        (task.pillarId ?? "").toLowerCase() === selectedPillar.toLowerCase();
       return matchesSearch && matchesPillar;
     });
   }, [normalizedSearch, selectedPillar, tasks]);
@@ -266,12 +345,14 @@ export default function TaskEditorPage() {
 
     const entries = filteredTasks.map((task) => ({
       task,
-      pillarName: pillarNamesById.get(task.pillarId ?? '') ?? '',
+      pillarName: pillarNamesById.get(task.pillarId ?? "") ?? "",
     }));
 
     entries.sort((a, b) => {
       if (a.pillarName === b.pillarName) {
-        return a.task.title.localeCompare(b.task.title, activeLocale, { sensitivity: 'base' });
+        return a.task.title.localeCompare(b.task.title, activeLocale, {
+          sensitivity: "base",
+        });
       }
 
       if (!a.pillarName) {
@@ -282,19 +363,28 @@ export default function TaskEditorPage() {
         return -1;
       }
 
-      return a.pillarName.localeCompare(b.pillarName, activeLocale, { sensitivity: 'base' });
+      return a.pillarName.localeCompare(b.pillarName, activeLocale, {
+        sensitivity: "base",
+      });
     });
 
     return entries.map((entry) => entry.task);
   }, [activeLocale, filteredTasks, pillarNamesById]);
 
-  const hasActiveFilters = normalizedSearch.length > 0 || selectedPillar.length > 0;
-  const isDeletingTask = deleteStatus === 'loading';
+  const hasActiveFilters =
+    normalizedSearch.length > 0 || selectedPillar.length > 0;
+  const isDeletingTask = deleteStatus === "loading";
 
-  const isTaskListEmpty = !isLoadingTasks && !combinedError && tasks.length === 0;
-  const visibleTasks = FEATURE_TASK_EDITOR_MOBILE_LIST_V1 ? sortedTasks : filteredTasks;
+  const isTaskListEmpty =
+    !isLoadingTasks && !combinedError && tasks.length === 0;
+  const visibleTasks = FEATURE_TASK_EDITOR_MOBILE_LIST_V1
+    ? sortedTasks
+    : filteredTasks;
   const isFilteredEmpty =
-    !isLoadingTasks && !combinedError && tasks.length > 0 && visibleTasks.length === 0;
+    !isLoadingTasks &&
+    !combinedError &&
+    tasks.length > 0 &&
+    visibleTasks.length === 0;
 
   const handleRetry = () => {
     reloadProfile();
@@ -305,6 +395,29 @@ export default function TaskEditorPage() {
     setShowCreateModal(true);
   };
 
+  const handleGuideStepChange = useCallback((stepId: EditorGuideStepId) => {
+    setActiveGuideStepId(stepId);
+    setTaskToEdit(null);
+    setShowSuggestionsModal(false);
+
+    switch (stepId) {
+      case "modal-core":
+      case "modal-ai-thinking":
+        setShowCreateModal(true);
+        break;
+      case "modal-entry":
+      case "filters":
+      case "task-list":
+      case "suggestions":
+      case "wheel-core":
+      case "wheel-pillars":
+      case "wheel-traits":
+      default:
+        setShowCreateModal(false);
+        break;
+    }
+  }, []);
+
   const handleDeleteModalClose = useCallback(() => {
     if (isDeletingTask) {
       return;
@@ -314,12 +427,12 @@ export default function TaskEditorPage() {
 
   const handleConfirmDelete = useCallback(async () => {
     if (!taskToDelete) {
-      setDeleteErrorMessage(t('editor.validation.taskNotFound'));
+      setDeleteErrorMessage(t("editor.validation.taskNotFound"));
       return;
     }
 
     if (!backendUserId) {
-      setDeleteErrorMessage(t('editor.validation.userNotFound'));
+      setDeleteErrorMessage(t("editor.validation.userNotFound"));
       return;
     }
 
@@ -329,9 +442,10 @@ export default function TaskEditorPage() {
       await deleteTask(backendUserId, taskToDelete.id);
       setTaskToDelete(null);
     } catch (error) {
-      const message = error instanceof Error ? error.message : t('editor.toast.delete.error');
+      const message =
+        error instanceof Error ? error.message : t("editor.toast.delete.error");
       setDeleteErrorMessage(message);
-      setPageToast({ type: 'error', text: message });
+      setPageToast({ type: "error", text: message });
     }
   }, [backendUserId, deleteTask, t, taskToDelete]);
 
@@ -348,8 +462,8 @@ export default function TaskEditorPage() {
     async (task: UserTask) => {
       if (!backendUserId) {
         setPageToast({
-          type: 'error',
-          text: t('editor.validation.userNotFound'),
+          type: "error",
+          text: t("editor.validation.userNotFound"),
         });
         return;
       }
@@ -357,9 +471,12 @@ export default function TaskEditorPage() {
       setDuplicatingTaskId(task.id);
 
       try {
-        const normalizedTitle = task.title?.trim() ?? '';
-        const title = normalizedTitle.length > 0 ? `${normalizedTitle} ${t('editor.duplicate.copySuffix')}` : t('editor.duplicate.fallbackTitle');
-        await duplicateTask(backendUserId, {
+        const normalizedTitle = task.title?.trim() ?? "";
+        const title =
+          normalizedTitle.length > 0
+            ? `${normalizedTitle} ${t("editor.duplicate.copySuffix")}`
+            : t("editor.duplicate.fallbackTitle");
+        await createTask(backendUserId, {
           title,
           pillarId: task.pillarId ?? null,
           traitId: task.traitId ?? null,
@@ -367,34 +484,120 @@ export default function TaskEditorPage() {
           difficultyId: task.difficultyId ?? null,
           isActive: task.isActive ?? true,
         });
-        setPageToast({ type: 'success', text: t('editor.toast.duplicate.success') });
+        setPageToast({
+          type: "success",
+          text: t("editor.toast.duplicate.success"),
+        });
       } catch (error) {
-        const message = error instanceof Error ? error.message : t('editor.toast.duplicate.error');
-        setPageToast({ type: 'error', text: message });
+        const message =
+          error instanceof Error
+            ? error.message
+            : t("editor.toast.duplicate.error");
+        setPageToast({ type: "error", text: message });
       } finally {
         setDuplicatingTaskId(null);
       }
     },
-    [backendUserId, duplicateTask, t],
+    [backendUserId, createTask, t],
   );
 
-  const handleImproveTask = useCallback(
-    (task: UserTask) => {
-      if (!task) {
+  const handleApplySuggestions = useCallback(
+    async (selectedTaskIds: string[]) => {
+      if (!backendUserId) {
+        setPageToast({
+          type: "error",
+          text: t("editor.validation.userNotFound"),
+        });
+        return;
+      }
+      if (selectedTaskIds.length === 0) {
         return;
       }
 
-      // TODO: conectar con el flujo de mejora por IA cuando esté disponible en el editor.
-      setPageToast({
-        type: 'info',
-        text: t('editor.toast.improve.comingSoon'),
-      });
+      setIsApplyingSuggestions(true);
+      const selectedTasks = EDITOR_LAB_QUICK_START_SEED.filter((task) =>
+        selectedTaskIds.includes(task.id),
+      );
+
+      try {
+        const normalizeValue = (value: string) => value.trim().toLowerCase();
+        const pillarIdBySeedPillar = new Map<string, string>();
+        for (const pillar of pillars) {
+          const name = normalizeValue(pillar.name);
+          const code = normalizeValue(pillar.code ?? "");
+          if (
+            name.includes("body") ||
+            name.includes("cuerpo") ||
+            code.includes("body")
+          ) {
+            pillarIdBySeedPillar.set("Body", pillar.id);
+          }
+          if (
+            name.includes("mind") ||
+            name.includes("mente") ||
+            code.includes("mind")
+          ) {
+            pillarIdBySeedPillar.set("Mind", pillar.id);
+          }
+          if (
+            name.includes("soul") ||
+            name.includes("alma") ||
+            code.includes("soul")
+          ) {
+            pillarIdBySeedPillar.set("Soul", pillar.id);
+          }
+        }
+
+        const traitLookup = new Map<string, string>();
+        for (const [seedPillar, pillarId] of pillarIdBySeedPillar.entries()) {
+          const traits = await fetchCatalogTraits(pillarId);
+          for (const trait of traits) {
+            const key = `${seedPillar}:${normalizeValue(trait.code ?? trait.name)}`;
+            traitLookup.set(key, trait.id);
+          }
+        }
+
+        let createdCount = 0;
+        for (const task of selectedTasks) {
+          const pillarId = pillarIdBySeedPillar.get(task.pillar) ?? null;
+          const traitId =
+            traitLookup.get(`${task.pillar}:${normalizeValue(task.trait)}`) ??
+            null;
+          await createTask(backendUserId, {
+            title: task.title,
+            pillarId,
+            traitId,
+            statId: null,
+            difficultyId: null,
+            isActive: true,
+          });
+          createdCount += 1;
+        }
+
+        await reloadTasks();
+        setShowSuggestionsModal(false);
+        setPageToast({
+          type: "success",
+          text:
+            language === "es"
+              ? `Se agregaron ${createdCount} sugerencia(s) a tu sistema.`
+              : `Added ${createdCount} suggestion(s) to your system.`,
+        });
+      } catch (error) {
+        const message =
+          error instanceof Error
+            ? error.message
+            : "Could not apply suggestions.";
+        setPageToast({ type: "error", text: message });
+      } finally {
+        setIsApplyingSuggestions(false);
+      }
     },
-    [t],
+    [backendUserId, createTask, language, pillars, reloadTasks, t],
   );
 
   const navigationTasks = useMemo(() => {
-    if (editVariant !== 'panel') {
+    if (editVariant !== "panel") {
       return [] as UserTask[];
     }
 
@@ -403,12 +606,12 @@ export default function TaskEditorPage() {
 
   const handleCloseEdit = useCallback(() => {
     setTaskToEdit(null);
-    setEditVariant('modal');
+    setEditVariant("modal");
     setEditGroupKey(null);
   }, []);
 
   const scrollToEditorTop = useCallback(() => {
-    if (typeof window === 'undefined') {
+    if (typeof window === "undefined") {
       return;
     }
 
@@ -418,8 +621,9 @@ export default function TaskEditorPage() {
     }
 
     const headerOffset = 92;
-    const target = anchor.getBoundingClientRect().top + window.scrollY - headerOffset;
-    window.scrollTo({ top: Math.max(target, 0), behavior: 'smooth' });
+    const target =
+      anchor.getBoundingClientRect().top + window.scrollY - headerOffset;
+    window.scrollTo({ top: Math.max(target, 0), behavior: "smooth" });
   }, []);
 
   const handleEditSuccess = useCallback(() => {
@@ -433,8 +637,8 @@ export default function TaskEditorPage() {
 
     if (!isFirstOnboardingEdit) {
       setPageToast({
-        type: 'success',
-        text: t('editor.toast.edit.saved'),
+        type: "success",
+        text: t("editor.toast.edit.saved"),
       });
     }
 
@@ -453,7 +657,7 @@ export default function TaskEditorPage() {
 
   const handleNavigatePanelTask = useCallback(
     (taskId: string) => {
-      if (editVariant !== 'panel') {
+      if (editVariant !== "panel") {
         return;
       }
 
@@ -472,13 +676,16 @@ export default function TaskEditorPage() {
           title={activeSection.pageTitle}
           sections={sections.map((section) => ({
             ...section,
-            showPulseDot: section.key === 'dashboard' && shouldShowDashboardDot,
+            showPulseDot: section.key === "dashboard" && shouldShowDashboardDot,
           }))}
         />
-        <main className="flex-1 pb-[calc(env(safe-area-inset-bottom,0px)+10.25rem)] md:pb-0" data-light-scope="editor">
+        <main
+          className="flex-1 pb-[calc(env(safe-area-inset-bottom,0px)+10.25rem)] md:pb-0"
+          data-light-scope="editor"
+        >
           <div ref={editorTopRef} className="scroll-mt-24" />
           <div className="mx-auto w-full max-w-7xl px-2 py-4 md:px-5 md:py-6 lg:px-6 lg:py-8">
-            <SectionHeader section={taskEditorSection} />
+            <SectionHeader section={taskEditorSection} showLabBadge={location.pathname.startsWith("/labs/")} />
             <Card className="px-4 py-5 md:p-6">
               <div className="flex flex-col gap-5">
                 {pageToast && (
@@ -494,13 +701,22 @@ export default function TaskEditorPage() {
                   <div className="ib-onboarding-alert ib-onboarding-alert--progress mb-2 rounded-2xl px-3 py-2.5 md:mb-3 md:px-4">
                     <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
                       <p className="ib-onboarding-alert__body text-xs leading-snug md:text-sm">
-                        {t('editor.onboarding.banner.message')}
+                        {t("editor.onboarding.banner.message")}
                       </p>
                       <Link
-                        to={getDashboardSectionConfig('dashboard', location.pathname, language).to}
+                        to={
+                          getDashboardSectionConfig(
+                            "dashboard",
+                            location.pathname,
+                            language,
+                          ).to
+                        }
                         className="ib-onboarding-alert__cta inline-flex shrink-0 items-center justify-center rounded-full px-3 py-1.5 text-xs font-semibold tracking-wide transition focus-visible:outline-none focus-visible:ring-2"
                       >
-                        {t('editor.onboarding.banner.cta')} <span className="ml-1.5" aria-hidden>→</span>
+                        {t("editor.onboarding.banner.cta")}{" "}
+                        <span className="ml-1.5" aria-hidden>
+                          →
+                        </span>
                       </Link>
                     </div>
                   </div>
@@ -515,42 +731,58 @@ export default function TaskEditorPage() {
                   isLoadingPillars={isLoadingPillars}
                   pillarsError={pillarsError}
                   onRetryPillars={reloadPillars}
+                  onOpenGuide={() => setShowGuideModal(true)}
+                  onOpenSuggestions={() => setShowSuggestionsModal(true)}
                 />
 
                 {isLoadingTasks && <TaskListSkeleton />}
 
                 {combinedError && !isLoadingTasks && (
-                  <TaskListError message={combinedError.message} onRetry={handleRetry} />
+                  <TaskListError
+                    message={combinedError.message}
+                    onRetry={handleRetry}
+                  />
                 )}
 
                 {isTaskListEmpty && (
-                  <TaskListEmpty message={t('editor.empty.noTasks')} />
+                  <TaskListEmpty message={t("editor.empty.noTasks")} />
                 )}
 
                 {isFilteredEmpty && (
                   <TaskListEmpty
                     message={
                       hasActiveFilters
-                        ? t('editor.empty.noMatches')
-                        : t('editor.empty.noVisibleTasks')
+                        ? t("editor.empty.noMatches")
+                        : t("editor.empty.noVisibleTasks")
                     }
                   />
                 )}
 
-                {!isLoadingTasks && !combinedError && visibleTasks.length > 0 && (
-                  <TaskList
-                    tasks={visibleTasks}
-                    pillarNamesById={pillarNamesById}
-                    traitNamesById={traitNamesMap}
-                    statNamesById={statNamesMap}
-                    difficultyNamesById={difficultyNamesById}
-                    onEditTask={(task) => setTaskToEdit(task)}
-                    onDeleteTask={(task) => setTaskToDelete(task)}
-                    onDuplicateTask={FEATURE_TASK_EDITOR_MOBILE_LIST_V1 ? handleDuplicateTask : undefined}
-                    onImproveTask={FEATURE_TASK_EDITOR_MOBILE_LIST_V1 ? handleImproveTask : undefined}
-                    duplicatingTaskId={FEATURE_TASK_EDITOR_MOBILE_LIST_V1 ? duplicatingTaskId : null}
-                  />
-                )}
+                {!isLoadingTasks &&
+                  !combinedError &&
+                  visibleTasks.length > 0 && (
+                    <div data-editor-guide-target="task-list">
+                      <TaskList
+                        tasks={visibleTasks}
+                        pillarNamesById={pillarNamesById}
+                        traitNamesById={traitNamesMap}
+                        statNamesById={statNamesMap}
+                        difficultyNamesById={difficultyNamesById}
+                        onEditTask={(task) => setTaskToEdit(task)}
+                        onDeleteTask={(task) => setTaskToDelete(task)}
+                        onDuplicateTask={
+                          FEATURE_TASK_EDITOR_MOBILE_LIST_V1
+                            ? handleDuplicateTask
+                            : undefined
+                        }
+                        duplicatingTaskId={
+                          FEATURE_TASK_EDITOR_MOBILE_LIST_V1
+                            ? duplicatingTaskId
+                            : null
+                        }
+                      />
+                    </div>
+                  )}
               </div>
             </Card>
           </div>
@@ -562,11 +794,15 @@ export default function TaskEditorPage() {
 
               return {
                 key: section.key,
-                label: section.key === 'editor' ? t('editor.navigation.label') : section.label,
+                label:
+                  section.key === "editor"
+                    ? t("editor.navigation.label")
+                    : section.label,
                 to: section.to,
                 icon: <Icon className="h-4 w-4" />,
                 end: section.end,
-                showPulseDot: section.key === 'dashboard' && shouldShowDashboardDot,
+                showPulseDot:
+                  section.key === "dashboard" && shouldShowDashboardDot,
               };
             })}
           />
@@ -575,10 +811,13 @@ export default function TaskEditorPage() {
           <button
             type="button"
             onClick={handleCreateClick}
+            data-editor-guide-target="new-task-trigger"
             className="fixed bottom-[calc(env(safe-area-inset-bottom,0px)+7.25rem)] right-4 z-40 inline-flex items-center gap-1.5 rounded-full bg-violet-500 px-3.5 py-2 text-[0.72rem] font-semibold text-white shadow-[0_10px_26px_rgba(139,92,246,0.42)] transition hover:bg-violet-400 focus:outline-none focus-visible:ring-2 focus-visible:ring-violet-200 focus-visible:ring-offset-2 focus-visible:ring-offset-slate-950 md:bottom-10 md:right-8 md:gap-2 md:px-5 md:py-3 md:text-sm"
           >
-            <span aria-hidden className="text-sm leading-none md:text-lg">＋</span>
-            {t('editor.button.newTask')}
+            <span aria-hidden className="text-sm leading-none md:text-lg">
+              ＋
+            </span>
+            {t("editor.button.newTask")}
           </button>
         )}
         <CreateTaskModal
@@ -589,6 +828,7 @@ export default function TaskEditorPage() {
           isLoadingPillars={isLoadingPillars}
           pillarsError={pillarsError}
           onRetryPillars={reloadPillars}
+          guideStepId={showGuideModal ? activeGuideStepId : null}
         />
         <EditTaskModal
           open={taskToEdit != null}
@@ -609,14 +849,36 @@ export default function TaskEditorPage() {
           errorMessage={deleteErrorMessage}
           onConfirm={handleConfirmDelete}
         />
+        <EditorGuideOverlay
+          isOpen={showGuideModal}
+          locale={language === "es" ? "es" : "en"}
+          onStepChange={handleGuideStepChange}
+          onClose={() => {
+            markEditorGuideAsSeen();
+            setShowGuideModal(false);
+            setShowCreateModal(false);
+          }}
+        />
+        <SuggestionsLabModal
+          isOpen={showSuggestionsModal}
+          isApplying={isApplyingSuggestions}
+          onClose={() => setShowSuggestionsModal(false)}
+          onApply={handleApplySuggestions}
+        />
       </div>
     </DevErrorBoundary>
   );
 }
 
-function SectionHeader({ section }: { section: DashboardSectionConfig }) {
+function SectionHeader({
+  section,
+  showLabBadge = false,
+}: {
+  section: DashboardSectionConfig;
+  showLabBadge?: boolean;
+}) {
   const normalizedTitle = section.contentTitle.trim();
-  const normalizedDescription = section.description?.trim() ?? '';
+  const normalizedDescription = section.description?.trim() ?? "";
   const shouldShowTitle = normalizedTitle.length > 0;
   const shouldShowDescription = normalizedDescription.length > 0;
 
@@ -625,13 +887,22 @@ function SectionHeader({ section }: { section: DashboardSectionConfig }) {
   }
 
   return (
-    <header className="mb-6 space-y-2 md:mb-8">
+    <header className="mb-6 space-y-3 md:mb-8">
+      {showLabBadge && (
+        <p className="text-[10px] font-semibold uppercase tracking-[0.28em] text-violet-200/85">
+          Editor Lab
+        </p>
+      )}
       {shouldShowTitle && (
-        <h1 className="font-display text-2xl font-semibold text-[color:var(--color-text)] sm:text-3xl">
+        <h1 className="font-display text-2xl font-semibold tracking-[-0.02em] text-[color:var(--color-text)] sm:text-3xl">
           {normalizedTitle}
         </h1>
       )}
-      {shouldShowDescription && <p className="text-sm text-[color:var(--color-slate-400)]">{normalizedDescription}</p>}
+      {shouldShowDescription && (
+        <p className="max-w-3xl text-sm leading-relaxed text-[color:var(--color-slate-400)]">
+          {normalizedDescription}
+        </p>
+      )}
     </header>
   );
 }
@@ -645,6 +916,8 @@ interface TaskFiltersProps {
   isLoadingPillars: boolean;
   pillarsError: Error | null;
   onRetryPillars: () => void;
+  onOpenGuide: () => void;
+  onOpenSuggestions: () => void;
 }
 
 function TaskFilters({
@@ -656,28 +929,30 @@ function TaskFilters({
   isLoadingPillars,
   pillarsError,
   onRetryPillars,
+  onOpenGuide,
+  onOpenSuggestions,
 }: TaskFiltersProps) {
-  const { t } = usePostLoginLanguage();
+  const { t, language } = usePostLoginLanguage();
   if (!FEATURE_TASK_EDITOR_MOBILE_LIST_V1) {
     return (
       <div className="flex flex-col gap-3 md:flex-row md:items-end">
         <label className="flex w-full flex-col gap-2">
           <span className="text-xs font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-400)]">
-            {t('editor.filters.search.label')}
+            {t("editor.filters.search.label")}
           </span>
           <div className="relative flex items-center">
             <input
               type="search"
               value={searchTerm}
               onChange={(event) => onSearchChange(event.target.value)}
-              placeholder={t('editor.filters.search.placeholder')}
+              placeholder={t("editor.filters.search.placeholder")}
               className="w-full rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2.5 text-sm ios-touch-input text-[color:var(--color-slate-100)] placeholder:text-[color:var(--color-slate-400)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
             />
           </div>
         </label>
         <label className="flex w-full flex-col gap-2 md:max-w-xs">
           <span className="text-xs font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-400)]">
-            {t('editor.filters.pillar.label')}
+            {t("editor.filters.pillar.label")}
           </span>
           <select
             value={selectedPillar}
@@ -685,13 +960,19 @@ function TaskFilters({
             className="w-full appearance-none rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2.5 text-sm ios-touch-input text-[color:var(--color-slate-100)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
           >
             {pillars.map((pillar) => (
-              <option key={pillar.value || 'all'} value={pillar.value} className="bg-slate-900 text-[color:var(--color-slate-100)]">
+              <option
+                key={pillar.value || "all"}
+                value={pillar.value}
+                className="bg-slate-900 text-[color:var(--color-slate-100)]"
+              >
                 {pillar.label}
               </option>
             ))}
           </select>
           {isLoadingPillars && (
-            <span className="text-[11px] uppercase tracking-[0.2em] text-slate-500">{t('editor.filters.pillar.loading')}</span>
+            <span className="text-[11px] uppercase tracking-[0.2em] text-slate-500">
+              {t("editor.filters.pillar.loading")}
+            </span>
           )}
           {pillarsError && !isLoadingPillars && (
             <button
@@ -699,7 +980,7 @@ function TaskFilters({
               onClick={onRetryPillars}
               className="self-start text-[11px] font-semibold uppercase tracking-[0.2em] text-rose-300"
             >
-              {t('editor.filters.pillar.retry')}
+              {t("editor.filters.pillar.retry")}
             </button>
           )}
         </label>
@@ -709,24 +990,69 @@ function TaskFilters({
 
   return (
     <div className="flex flex-col gap-4">
+      <div className="hidden items-center justify-end gap-2 md:flex">
+        <button
+          type="button"
+          onClick={onOpenGuide}
+          aria-label={language === "es" ? "Abrir guía" : "Open guide"}
+          title={language === "es" ? "Ver guía" : "View guide"}
+          className="inline-flex items-center gap-1.5 rounded-full border border-violet-400/55 bg-violet-100/95 px-2.5 py-1 text-xs font-semibold text-violet-700 shadow-[0_4px_12px_rgba(124,58,237,0.16)] transition hover:border-violet-500/60 hover:bg-violet-200/90 hover:text-violet-800 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-violet-400 disabled:cursor-not-allowed disabled:opacity-45 dark:border-violet-300/55 dark:bg-violet-500/35 dark:text-violet-50 dark:shadow-[0_6px_18px_rgba(124,58,237,0.35)] dark:hover:border-violet-200/80 dark:hover:bg-violet-500/45 dark:hover:text-white dark:focus-visible:outline-violet-200"
+        >
+          <GuideCompassIcon className="h-3.5 w-3.5" />
+          <span>{language === "es" ? "Ver guía" : "View guide"}</span>
+        </button>
+        <button
+          type="button"
+          onClick={onOpenSuggestions}
+          aria-label={language === "es" ? "Abrir sugerencias" : "Open suggestions"}
+          data-editor-guide-target="suggestions"
+          className="inline-flex h-9 w-9 items-center justify-center rounded-full border border-violet-300/40 bg-[linear-gradient(170deg,rgba(167,139,250,0.26),rgba(129,140,248,0.08))] text-violet-100 shadow-[inset_0_1px_0_rgba(255,255,255,0.2),0_10px_18px_rgba(76,29,149,0.3)] transition hover:border-violet-200/70 hover:bg-violet-500/20 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-violet-200/65"
+        >
+          <SuggestionsMagicIcon className="h-[18px] w-[18px]" />
+        </button>
+      </div>
       <div className="md:hidden">
         <div className="editor-filters-mobile-panel sticky -mx-4 -mt-5 top-[4.5rem] z-30 space-y-3 rounded-t-2xl px-4 pt-5 pb-3 bg-[color:var(--color-slate-900-95)]">
-          <label className="flex flex-col gap-1">
-            <input
-              type="search"
-              value={searchTerm}
-              onChange={(event) => onSearchChange(event.target.value)}
-              placeholder={t('editor.filters.search.placeholder')}
-              className="w-full rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2 text-sm ios-touch-input text-[color:var(--color-slate-100)] placeholder:text-[color:var(--color-slate-400)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
-            />
-          </label>
-          <div className="space-y-2">
+          <div className="flex items-center justify-end gap-2">
+            <button
+              type="button"
+              onClick={onOpenGuide}
+              aria-label={language === "es" ? "Abrir guía" : "Open guide"}
+              title={language === "es" ? "Ver guía" : "View guide"}
+              className="inline-flex items-center gap-1.5 rounded-full border border-violet-400/55 bg-violet-100/95 px-2.5 py-1 text-xs font-semibold text-violet-700 shadow-[0_4px_12px_rgba(124,58,237,0.16)] transition hover:border-violet-500/60 hover:bg-violet-200/90 hover:text-violet-800 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-violet-400 disabled:cursor-not-allowed disabled:opacity-45 dark:border-violet-300/55 dark:bg-violet-500/35 dark:text-violet-50 dark:shadow-[0_6px_18px_rgba(124,58,237,0.35)] dark:hover:border-violet-200/80 dark:hover:bg-violet-500/45 dark:hover:text-white dark:focus-visible:outline-violet-200"
+            >
+              <GuideCompassIcon className="h-3.5 w-3.5" />
+              <span>{language === "es" ? "Ver guía" : "View guide"}</span>
+            </button>
+            <button
+              type="button"
+              onClick={onOpenSuggestions}
+              aria-label={language === "es" ? "Abrir sugerencias" : "Open suggestions"}
+              data-editor-guide-target="suggestions"
+              className="inline-flex h-9 w-9 items-center justify-center rounded-full border border-violet-300/40 bg-[linear-gradient(170deg,rgba(167,139,250,0.26),rgba(129,140,248,0.08))] text-violet-100 shadow-[inset_0_1px_0_rgba(255,255,255,0.2),0_10px_18px_rgba(76,29,149,0.3)] transition hover:border-violet-200/70 hover:bg-violet-500/20"
+            >
+              <SuggestionsMagicIcon className="h-[18px] w-[18px]" />
+            </button>
+          </div>
+          <div
+            data-editor-guide-target="search-pillar-zone"
+            className="space-y-2 rounded-2xl border border-white/10 bg-[linear-gradient(180deg,rgba(15,23,42,0.48),rgba(30,41,59,0.28))] p-2.5 shadow-[inset_0_1px_0_rgba(255,255,255,0.08)]"
+          >
+            <label className="flex flex-col gap-1">
+              <input
+                type="search"
+                value={searchTerm}
+                onChange={(event) => onSearchChange(event.target.value)}
+                placeholder={t("editor.filters.search.placeholder")}
+                className="w-full rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2 text-sm ios-touch-input text-[color:var(--color-slate-100)] placeholder:text-[color:var(--color-slate-400)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
+              />
+            </label>
             <div className="flex gap-2 overflow-x-auto pb-1 [mask-image:linear-gradient(to_right,transparent,black_12%,black_88%,transparent)]">
               {pillars.map((pillar) => {
                 const isActive = pillar.value === selectedPillar;
                 return (
                   <button
-                    key={pillar.value || 'all'}
+                    key={pillar.value || "all"}
                     type="button"
                     onClick={() => onPillarChange(pillar.value)}
                     className="editor-pillar-chip inline-flex items-center whitespace-nowrap rounded-full border px-3 py-1 text-[10px] font-semibold uppercase tracking-[0.24em] transition focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white/60"
@@ -741,7 +1067,9 @@ function TaskFilters({
           </div>
         </div>
         {isLoadingPillars && (
-          <p className="px-1 text-[10px] uppercase tracking-[0.28em] text-slate-500">{t('editor.filters.pillar.loading')}</p>
+          <p className="px-1 text-[10px] uppercase tracking-[0.28em] text-slate-500">
+            {t("editor.filters.pillar.loading")}
+          </p>
         )}
         {pillarsError && !isLoadingPillars && (
           <button
@@ -749,28 +1077,31 @@ function TaskFilters({
             onClick={onRetryPillars}
             className="px-1 text-[10px] font-semibold uppercase tracking-[0.28em] text-rose-300"
           >
-            {t('editor.filters.pillar.retry')}
+            {t("editor.filters.pillar.retry")}
           </button>
         )}
       </div>
-      <div className="hidden flex-col gap-3 md:flex md:flex-row md:items-end">
+      <div
+        data-editor-guide-target="search-pillar-zone"
+        className="hidden flex-col gap-3 rounded-2xl border border-white/10 bg-white/[0.02] p-3 md:flex md:flex-row md:items-end"
+      >
         <label className="flex w-full flex-col gap-2">
           <span className="text-xs font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-400)]">
-            {t('editor.filters.search.label')}
+            {t("editor.filters.search.label")}
           </span>
           <div className="relative flex items-center">
             <input
               type="search"
               value={searchTerm}
               onChange={(event) => onSearchChange(event.target.value)}
-              placeholder={t('editor.filters.search.placeholder')}
+              placeholder={t("editor.filters.search.placeholder")}
               className="w-full rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2.5 text-sm ios-touch-input text-[color:var(--color-slate-100)] placeholder:text-[color:var(--color-slate-400)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
             />
           </div>
         </label>
         <label className="flex w-full flex-col gap-2 md:max-w-xs">
           <span className="text-xs font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-400)]">
-            {t('editor.filters.pillar.label')}
+            {t("editor.filters.pillar.label")}
           </span>
           <select
             value={selectedPillar}
@@ -778,13 +1109,19 @@ function TaskFilters({
             className="w-full appearance-none rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] px-4 py-2.5 text-sm ios-touch-input text-[color:var(--color-slate-100)] focus:border-[color:var(--color-border-soft)] focus:outline-none focus:ring-2 focus:ring-white/20"
           >
             {pillars.map((pillar) => (
-              <option key={pillar.value || 'all'} value={pillar.value} className="bg-slate-900 text-[color:var(--color-slate-100)]">
+              <option
+                key={pillar.value || "all"}
+                value={pillar.value}
+                className="bg-slate-900 text-[color:var(--color-slate-100)]"
+              >
                 {pillar.label}
               </option>
             ))}
           </select>
           {isLoadingPillars && (
-            <span className="text-[11px] uppercase tracking-[0.2em] text-slate-500">{t('editor.filters.pillar.loading')}</span>
+            <span className="text-[11px] uppercase tracking-[0.2em] text-slate-500">
+              {t("editor.filters.pillar.loading")}
+            </span>
           )}
           {pillarsError && !isLoadingPillars && (
             <button
@@ -792,7 +1129,7 @@ function TaskFilters({
               onClick={onRetryPillars}
               className="self-start text-[11px] font-semibold uppercase tracking-[0.2em] text-rose-300"
             >
-              {t('editor.filters.pillar.retry')}
+              {t("editor.filters.pillar.retry")}
             </button>
           )}
         </label>
@@ -810,7 +1147,7 @@ function TaskList({
   onEditTask,
   onDeleteTask,
   onDuplicateTask,
-  onImproveTask,
+  onImproveTask: _onImproveTask,
   duplicatingTaskId = null,
 }: {
   tasks: UserTask[];
@@ -831,10 +1168,18 @@ function TaskList({
           <TaskCard
             key={task.id}
             task={task}
-            pillarName={pillarNamesById.get(task.pillarId ?? '') ?? null}
-            traitName={task.traitId ? traitNamesById.get(task.traitId) ?? null : null}
-            statName={task.statId ? statNamesById.get(task.statId) ?? null : null}
-            difficultyName={task.difficultyId ? difficultyNamesById.get(task.difficultyId) ?? null : null}
+            pillarName={pillarNamesById.get(task.pillarId ?? "") ?? null}
+            traitName={
+              task.traitId ? (traitNamesById.get(task.traitId) ?? null) : null
+            }
+            statName={
+              task.statId ? (statNamesById.get(task.statId) ?? null) : null
+            }
+            difficultyName={
+              task.difficultyId
+                ? (difficultyNamesById.get(task.difficultyId) ?? null)
+                : null
+            }
             onEdit={() => onEditTask(task)}
             onDelete={() => onDeleteTask(task)}
           />
@@ -848,10 +1193,10 @@ function TaskList({
       <div className="md:hidden">
         <TaskListMobile
           tasks={tasks}
+          pillarNamesById={pillarNamesById}
           onEditTask={onEditTask}
           onDeleteTask={onDeleteTask}
           onDuplicateTask={onDuplicateTask}
-          onImproveTask={onImproveTask}
           duplicatingTaskId={duplicatingTaskId}
         />
       </div>
@@ -860,10 +1205,18 @@ function TaskList({
           <TaskCard
             key={task.id}
             task={task}
-            pillarName={pillarNamesById.get(task.pillarId ?? '') ?? null}
-            traitName={task.traitId ? traitNamesById.get(task.traitId) ?? null : null}
-            statName={task.statId ? statNamesById.get(task.statId) ?? null : null}
-            difficultyName={task.difficultyId ? difficultyNamesById.get(task.difficultyId) ?? null : null}
+            pillarName={pillarNamesById.get(task.pillarId ?? "") ?? null}
+            traitName={
+              task.traitId ? (traitNamesById.get(task.traitId) ?? null) : null
+            }
+            statName={
+              task.statId ? (statNamesById.get(task.statId) ?? null) : null
+            }
+            difficultyName={
+              task.difficultyId
+                ? (difficultyNamesById.get(task.difficultyId) ?? null)
+                : null
+            }
             onEdit={() => onEditTask(task)}
             onDelete={() => onDeleteTask(task)}
           />
@@ -875,17 +1228,17 @@ function TaskList({
 
 function TaskListMobile({
   tasks,
+  pillarNamesById,
   onEditTask,
   onDeleteTask,
   onDuplicateTask,
-  onImproveTask,
   duplicatingTaskId,
 }: {
   tasks: UserTask[];
+  pillarNamesById: Map<string, string>;
   onEditTask: (task: UserTask) => void;
   onDeleteTask: (task: UserTask) => void;
   onDuplicateTask?: (task: UserTask) => void;
-  onImproveTask?: (task: UserTask) => void;
   duplicatingTaskId: string | null;
 }) {
   const { t } = usePostLoginLanguage();
@@ -909,16 +1262,16 @@ function TaskListMobile({
     };
 
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
+      if (event.key === "Escape") {
         setOpenMenuTaskId(null);
       }
     };
 
-    window.addEventListener('pointerdown', handlePointerDown);
-    window.addEventListener('keydown', handleKeyDown);
+    window.addEventListener("pointerdown", handlePointerDown);
+    window.addEventListener("keydown", handleKeyDown);
     return () => {
-      window.removeEventListener('pointerdown', handlePointerDown);
-      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener("pointerdown", handlePointerDown);
+      window.removeEventListener("keydown", handleKeyDown);
     };
   }, [openMenuTaskId]);
 
@@ -934,6 +1287,7 @@ function TaskListMobile({
       {tasks.map((task) => {
         const isMenuOpen = openMenuTaskId === task.id;
         const isDuplicating = duplicatingTaskId === task.id;
+        const pillarName = pillarNamesById.get(task.pillarId ?? "") ?? null;
 
         return (
           <li key={task.id} className="relative">
@@ -943,8 +1297,17 @@ function TaskListMobile({
               className="flex w-full flex-col gap-2 px-3 py-2.5 text-left transition hover:bg-[color:var(--color-overlay-2)] focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white/60"
             >
               <div className="flex items-start justify-between gap-3">
-                <p className="line-clamp-1 pr-8 text-sm font-semibold text-white">{task.title}</p>
+                <p className="line-clamp-1 pr-8 text-sm font-semibold text-white">
+                  {task.title}
+                </p>
               </div>
+              {pillarName && (
+                <div className="flex flex-wrap items-center gap-2 text-[11px] text-[color:var(--color-slate-400)]">
+                  <span className="inline-flex items-center gap-1 rounded-full border border-white/10 bg-white/5 px-1.5 py-px text-[9px] font-medium uppercase tracking-[0.1em] text-[color:var(--color-slate-400)]">
+                    {pillarName}
+                  </span>
+                </div>
+              )}
             </button>
             <div
               className="pointer-events-auto absolute right-1.5 bottom-2"
@@ -960,12 +1323,14 @@ function TaskListMobile({
                 aria-expanded={isMenuOpen}
                 onClick={(event) => {
                   event.stopPropagation();
-                  setOpenMenuTaskId((current) => (current === task.id ? null : task.id));
+                  setOpenMenuTaskId((current) =>
+                    current === task.id ? null : task.id,
+                  );
                 }}
                 className="flex h-8 w-8 items-center justify-center rounded-full border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] text-base text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:bg-[color:var(--color-overlay-2)] focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white/60"
               >
                 <span aria-hidden>⋯</span>
-                <span className="sr-only">{t('editor.task.actions.more')}</span>
+                <span className="sr-only">{t("editor.task.actions.more")}</span>
               </button>
               {isMenuOpen && (
                 <div className="absolute right-0 top-10 z-40 w-44 rounded-xl border border-[color:var(--color-border-subtle)] bg-[color:var(--color-slate-900-95)] p-1 shadow-[0_10px_30px_rgba(15,23,42,0.6)]">
@@ -978,7 +1343,7 @@ function TaskListMobile({
                     }}
                     className="flex w-full items-center gap-2 rounded-lg px-3 py-2 text-left text-sm text-[color:var(--color-slate-100)] transition hover:bg-[color:var(--color-overlay-2)]"
                   >
-                    {t('editor.button.edit')}
+                    {t("editor.button.edit")}
                   </button>
                   <button
                     type="button"
@@ -992,29 +1357,13 @@ function TaskListMobile({
                     }}
                     className={`flex w-full items-center gap-2 rounded-lg px-3 py-2 text-left text-sm transition ${
                       !onDuplicateTask
-                        ? 'cursor-not-allowed text-slate-500'
-                        : 'text-[color:var(--color-slate-100)] hover:bg-[color:var(--color-overlay-2)]'
-                    } ${isDuplicating ? 'opacity-70' : ''}`.trim()}
+                        ? "cursor-not-allowed text-slate-500"
+                        : "text-[color:var(--color-slate-100)] hover:bg-[color:var(--color-overlay-2)]"
+                    } ${isDuplicating ? "opacity-70" : ""}`.trim()}
                   >
-                    {isDuplicating ? t('editor.button.duplicating') : t('editor.button.duplicate')}
-                  </button>
-                  <button
-                    type="button"
-                    disabled={!onImproveTask}
-                    onClick={(event) => {
-                      event.stopPropagation();
-                      setOpenMenuTaskId(null);
-                      if (onImproveTask) {
-                        onImproveTask(task);
-                      }
-                    }}
-                    className={`flex w-full items-center gap-2 rounded-lg px-3 py-2 text-left text-sm transition ${
-                      onImproveTask
-                        ? 'text-[color:var(--color-slate-100)] hover:bg-[color:var(--color-overlay-2)]'
-                        : 'cursor-not-allowed text-slate-500'
-                    }`}
-                  >
-                    {t('editor.button.improveAi')}
+                    {isDuplicating
+                      ? t("editor.button.duplicating")
+                      : t("editor.button.duplicate")}
                   </button>
                   <button
                     type="button"
@@ -1024,7 +1373,9 @@ function TaskListMobile({
                       onDeleteTask(task);
                     }}
                     className="flex w-full items-center gap-2 rounded-lg px-3 py-2 text-left text-sm text-rose-200 transition hover:bg-rose-500/10"
-                  >{t('editor.modal.delete.confirm')}</button>
+                  >
+                    {t("editor.modal.delete.confirm")}
+                  </button>
                 </div>
               )}
             </div>
@@ -1057,55 +1408,82 @@ function TaskCard({
   return (
     <article className="group relative flex flex-col gap-3 rounded-2xl border border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] p-4 shadow-[0_8px_24px_rgba(15,23,42,0.35)] transition hover:border-[color:var(--color-border-soft)]">
       <div className="flex flex-col gap-3 sm:flex-row sm:items-start sm:justify-between">
-        <h3 className="font-semibold text-[color:var(--color-slate-100)]">{task.title}</h3>
+        <h3 className="font-semibold text-[color:var(--color-slate-100)]">
+          {task.title}
+        </h3>
         <div className="flex w-full flex-wrap items-center gap-2 sm:w-auto sm:justify-end">
           <span
             className={`rounded-full px-2.5 py-1 text-xs font-semibold uppercase tracking-[0.18em] ${
               task.isActive
-                ? 'bg-emerald-500/15 text-emerald-300'
-                : 'bg-slate-500/20 text-[color:var(--color-slate-300)]'
+                ? "bg-emerald-500/15 text-emerald-300"
+                : "bg-slate-500/20 text-[color:var(--color-slate-300)]"
             }`}
           >
-            {task.isActive ? t('editor.task.status.active') : t('editor.task.status.inactive')}
+            {task.isActive
+              ? t("editor.task.status.active")
+              : t("editor.task.status.inactive")}
           </span>
           <button
             type="button"
             onClick={onEdit}
             className="rounded-full border border-white/15 px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:text-white"
           >
-            {t('editor.button.edit')}
+            {t("editor.button.edit")}
           </button>
           <button
             type="button"
             onClick={onDelete}
             className="rounded-full border border-rose-500/30 px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] text-rose-200 transition hover:border-rose-400 hover:text-rose-100"
-          >{t('editor.modal.delete.confirm')}</button>
+          >
+            {t("editor.modal.delete.confirm")}
+          </button>
         </div>
       </div>
       <dl className="grid gap-1 text-xs text-[color:var(--color-slate-400)]">
         <div className="flex items-center justify-between gap-4">
-          <dt className="font-medium text-[color:var(--color-slate-300)]">{t('editor.field.pillar')}</dt>
-          <dd className="truncate text-right text-[color:var(--color-slate-200)]">{pillarName ?? task.pillarId ?? '—'}</dd>
+          <dt className="font-medium text-[color:var(--color-slate-300)]">
+            {t("editor.field.pillar")}
+          </dt>
+          <dd className="truncate text-right text-[color:var(--color-slate-200)]">
+            {pillarName ?? task.pillarId ?? "—"}
+          </dd>
         </div>
         <div className="flex items-center justify-between gap-4">
-          <dt className="font-medium text-[color:var(--color-slate-300)]">{t('editor.field.trait')}</dt>
-          <dd className="truncate text-right">{traitName ?? task.traitId ?? '—'}</dd>
+          <dt className="font-medium text-[color:var(--color-slate-300)]">
+            {t("editor.field.trait")}
+          </dt>
+          <dd className="truncate text-right">
+            {traitName ?? task.traitId ?? "—"}
+          </dd>
         </div>
         <div className="flex items-center justify-between gap-4">
-          <dt className="font-medium text-[color:var(--color-slate-300)]">{t('editor.field.stat')}</dt>
-          <dd className="truncate text-right">{statName ?? task.statId ?? '—'}</dd>
+          <dt className="font-medium text-[color:var(--color-slate-300)]">
+            {t("editor.field.stat")}
+          </dt>
+          <dd className="truncate text-right">
+            {statName ?? task.statId ?? "—"}
+          </dd>
         </div>
         <div className="flex items-center justify-between gap-4">
-          <dt className="font-medium text-[color:var(--color-slate-300)]">{t('editor.field.difficulty')}</dt>
-          <dd className="truncate text-right">{difficultyName ?? task.difficultyId ?? '—'}</dd>
+          <dt className="font-medium text-[color:var(--color-slate-300)]">
+            {t("editor.field.difficulty")}
+          </dt>
+          <dd className="truncate text-right">
+            {difficultyName ?? task.difficultyId ?? "—"}
+          </dd>
         </div>
         <div className="flex items-center justify-between gap-4">
-          <dt className="font-medium text-[color:var(--color-slate-300)]">{t('editor.field.baseGp')}</dt>
-          <dd className="truncate text-right">{task.xp != null ? task.xp : '—'}</dd>
+          <dt className="font-medium text-[color:var(--color-slate-300)]">
+            {t("editor.field.baseGp")}
+          </dt>
+          <dd className="truncate text-right">
+            {task.xp != null ? task.xp : "—"}
+          </dd>
         </div>
       </dl>
       <p className="text-[11px] uppercase tracking-[0.2em] text-slate-500">
-        {t('editor.field.updatedAt')}: {formatDateLabel(task.updatedAt, language)}
+        {t("editor.field.updatedAt")}:{" "}
+        {formatDateLabel(task.updatedAt, language)}
       </p>
     </article>
   );
@@ -1127,37 +1505,45 @@ interface TaskBoardProps {
   onDeleteTask: (task: UserTask) => void;
 }
 
-const PILLAR_STYLE_MAP: Record<string, { headerText: string; badgeBg: string; badgeText: string; bullet: string; ring: string }>
- = {
+const PILLAR_STYLE_MAP: Record<
+  string,
+  {
+    headerText: string;
+    badgeBg: string;
+    badgeText: string;
+    bullet: string;
+    ring: string;
+  }
+> = {
   BODY: {
-    headerText: 'text-emerald-300',
-    badgeBg: 'bg-emerald-500/15 border border-emerald-400/40',
-    badgeText: 'text-emerald-100',
-    bullet: 'text-emerald-300',
-    ring: 'ring-emerald-400/40',
+    headerText: "text-emerald-300",
+    badgeBg: "bg-emerald-500/15 border border-emerald-400/40",
+    badgeText: "text-emerald-100",
+    bullet: "text-emerald-300",
+    ring: "ring-emerald-400/40",
   },
   MIND: {
-    headerText: 'text-sky-300',
-    badgeBg: 'bg-sky-500/15 border border-sky-400/40',
-    badgeText: 'text-sky-100',
-    bullet: 'text-sky-300',
-    ring: 'ring-sky-400/40',
+    headerText: "text-sky-300",
+    badgeBg: "bg-sky-500/15 border border-sky-400/40",
+    badgeText: "text-sky-100",
+    bullet: "text-sky-300",
+    ring: "ring-sky-400/40",
   },
   SOUL: {
-    headerText: 'text-violet-300',
-    badgeBg: 'bg-violet-500/15 border border-violet-400/40',
-    badgeText: 'text-violet-100',
-    bullet: 'text-violet-300',
-    ring: 'ring-violet-400/40',
+    headerText: "text-violet-300",
+    badgeBg: "bg-violet-500/15 border border-violet-400/40",
+    badgeText: "text-violet-100",
+    bullet: "text-violet-300",
+    ring: "ring-violet-400/40",
   },
 };
 
 const DEFAULT_PILLAR_STYLE = {
-  headerText: 'text-indigo-300',
-  badgeBg: 'bg-indigo-500/15 border border-indigo-400/40',
-  badgeText: 'text-indigo-100',
-  bullet: 'text-indigo-300',
-  ring: 'ring-indigo-400/40',
+  headerText: "text-indigo-300",
+  badgeBg: "bg-indigo-500/15 border border-indigo-400/40",
+  badgeText: "text-indigo-100",
+  bullet: "text-indigo-300",
+  ring: "ring-indigo-400/40",
 };
 
 function resolvePillarStyle(code: string | undefined) {
@@ -1185,7 +1571,8 @@ function TaskBoard({
     <div className="hidden gap-4 lg:grid lg:grid-cols-3">
       {groups.map((group) => {
         const style = resolvePillarStyle(group.code);
-        const displayCode = group.code === 'UNKNOWN' ? t('editor.field.noPillar') : group.code;
+        const displayCode =
+          group.code === "UNKNOWN" ? t("editor.field.noPillar") : group.code;
 
         return (
           <section
@@ -1194,10 +1581,14 @@ function TaskBoard({
           >
             <header className="flex items-center justify-between border-b border-white/5 pb-3">
               <div className="space-y-1">
-                <p className={`text-xs font-semibold uppercase tracking-[0.24em] ${style.headerText}`}>
+                <p
+                  className={`text-xs font-semibold uppercase tracking-[0.24em] ${style.headerText}`}
+                >
                   {group.name}
                 </p>
-                <p className="text-[11px] uppercase tracking-[0.2em] text-slate-500">{displayCode}</p>
+                <p className="text-[11px] uppercase tracking-[0.2em] text-slate-500">
+                  {displayCode}
+                </p>
               </div>
               <span
                 className={`inline-flex items-center justify-center rounded-full px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] ${style.badgeBg} ${style.badgeText}`}
@@ -1208,7 +1599,7 @@ function TaskBoard({
             <div className="mt-3 flex-1 space-y-2">
               {group.tasks.length === 0 ? (
                 <p className="rounded-xl border border-white/5 bg-[color:var(--color-overlay-1)] px-3 py-6 text-center text-xs text-slate-500">
-                  {t('editor.board.emptyPillar')}
+                  {t("editor.board.emptyPillar")}
                 </p>
               ) : (
                 group.tasks.map((task: UserTask) => (
@@ -1216,11 +1607,17 @@ function TaskBoard({
                     key={task.id}
                     task={task}
                     groupKey={group.key}
-                    pillarName=
-                      {group.key === '__unknown__'
-                        ? pillarNamesById.get(task.pillarId ?? '') ?? t('editor.field.noPillar')
-                        : group.name}
-                    difficultyName={task.difficultyId ? difficultyNamesById.get(task.difficultyId) ?? null : null}
+                    pillarName={
+                      group.key === "__unknown__"
+                        ? (pillarNamesById.get(task.pillarId ?? "") ??
+                          t("editor.field.noPillar"))
+                        : group.name
+                    }
+                    difficultyName={
+                      task.difficultyId
+                        ? (difficultyNamesById.get(task.difficultyId) ?? null)
+                        : null
+                    }
                     isActiveTask={activeTaskId === task.id}
                     onSelectTask={onSelectTask}
                     onDeleteTask={onDeleteTask}
@@ -1267,18 +1664,18 @@ function TaskBoardItem({
   };
 
   const handleKeyDown = (event: ReactKeyboardEvent<HTMLDivElement>) => {
-    if (event.key === 'Enter' || event.key === ' ') {
+    if (event.key === "Enter" || event.key === " ") {
       event.preventDefault();
       onSelectTask(task, groupKey);
     }
   };
 
   const containerClasses = [
-    'group relative cursor-pointer rounded-xl border border-[color:var(--color-border-subtle)] bg-slate-900/60 p-3 transition hover:border-white/25 hover:bg-slate-900/70 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/30',
-    isActiveTask ? `border-white/30 bg-slate-900/80 ring-2 ${ringClass}` : '',
+    "group relative cursor-pointer rounded-xl border border-[color:var(--color-border-subtle)] bg-slate-900/60 p-3 transition hover:border-white/25 hover:bg-slate-900/70 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/30",
+    isActiveTask ? `border-white/30 bg-slate-900/80 ring-2 ${ringClass}` : "",
   ]
     .filter(Boolean)
-    .join(' ');
+    .join(" ");
 
   return (
     <div
@@ -1291,18 +1688,24 @@ function TaskBoardItem({
     >
       <div className="flex items-start justify-between gap-2">
         <div className="space-y-2">
-          <p className="text-sm font-semibold text-[color:var(--color-slate-100)]">{task.title}</p>
+          <p className="text-sm font-semibold text-[color:var(--color-slate-100)]">
+            {task.title}
+          </p>
           <div className="flex flex-wrap items-center gap-2 text-[11px] font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-400)]">
             <span
               className={`rounded-full px-2 py-0.5 text-[10px] uppercase tracking-[0.2em] ${
-                task.isActive ? 'bg-emerald-500/10 text-emerald-200' : 'bg-slate-700/20 text-[color:var(--color-slate-300)]'
+                task.isActive
+                  ? "bg-emerald-500/10 text-emerald-200"
+                  : "bg-slate-700/20 text-[color:var(--color-slate-300)]"
               }`}
             >
-              {task.isActive ? t('editor.task.status.active') : t('editor.task.status.inactive')}
+              {task.isActive
+                ? t("editor.task.status.active")
+                : t("editor.task.status.inactive")}
             </span>
             <span className="flex items-center gap-1 text-[color:var(--color-slate-400)]">
               <span className={`text-base leading-none ${bulletClass}`}>•</span>
-              {difficultyName ?? t('editor.field.noDifficulty')}
+              {difficultyName ?? t("editor.field.noDifficulty")}
             </span>
             <span className="text-slate-500">{pillarName}</span>
           </div>
@@ -1314,7 +1717,9 @@ function TaskBoardItem({
             onDeleteTask(task);
           }}
           className="rounded-full border border-rose-500/40 px-2 py-1 text-[10px] font-semibold uppercase tracking-[0.2em] text-rose-200 transition hover:border-rose-400 hover:text-rose-100 focus:outline-none focus-visible:ring-2 focus-visible:ring-rose-400/50"
-        >{t('editor.modal.delete.confirm')}</button>
+        >
+          {t("editor.modal.delete.confirm")}
+        </button>
       </div>
     </div>
   );
@@ -1329,7 +1734,14 @@ interface DeleteTaskModalProps {
   onConfirm: () => Promise<void>;
 }
 
-function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConfirm }: DeleteTaskModalProps) {
+function DeleteTaskModal({
+  open,
+  onClose,
+  task,
+  isDeleting,
+  errorMessage,
+  onConfirm,
+}: DeleteTaskModalProps) {
   const { t } = usePostLoginLanguage();
   useEffect(() => {
     if (!open) {
@@ -1337,15 +1749,15 @@ function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConf
     }
 
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape' && !isDeleting) {
+      if (event.key === "Escape" && !isDeleting) {
         event.preventDefault();
         onClose();
       }
     };
 
-    window.addEventListener('keydown', handleKeyDown);
+    window.addEventListener("keydown", handleKeyDown);
     return () => {
-      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener("keydown", handleKeyDown);
     };
   }, [open, isDeleting, onClose]);
 
@@ -1360,14 +1772,17 @@ function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConf
     await onConfirm();
   };
 
-  const normalizedTitle = task?.title?.trim() ?? '';
-  const displayTitle = normalizedTitle.length > 0 ? `“${normalizedTitle}”` : t('editor.modal.delete.thisTask');
+  const normalizedTitle = task?.title?.trim() ?? "";
+  const displayTitle =
+    normalizedTitle.length > 0
+      ? `“${normalizedTitle}”`
+      : t("editor.modal.delete.thisTask");
 
   return (
     <div className="fixed inset-0 z-[60] flex items-end justify-center bg-slate-950/70 backdrop-blur-sm md:items-center">
       <button
         type="button"
-        aria-label={t('editor.button.close')}
+        aria-label={t("editor.button.close")}
         onClick={() => {
           if (!isDeleting) {
             onClose();
@@ -1383,11 +1798,15 @@ function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConf
           onClick={(event) => event.stopPropagation()}
         >
           <header className="space-y-1">
-            <p className="text-[11px] font-semibold uppercase tracking-[0.24em] text-[color:var(--color-slate-400)]">{t('editor.modal.delete.heading')}</p>
-            <h2 className="text-xl font-semibold text-white">{t('editor.modal.delete.title')}</h2>
+            <p className="text-[11px] font-semibold uppercase tracking-[0.24em] text-[color:var(--color-slate-400)]">
+              {t("editor.modal.delete.heading")}
+            </p>
+            <h2 className="text-xl font-semibold text-white">
+              {t("editor.modal.delete.title")}
+            </h2>
           </header>
           <p className="text-sm text-[color:var(--color-slate-300)]">
-            {t('editor.modal.delete.message', { title: displayTitle })}
+            {t("editor.modal.delete.message", { title: displayTitle })}
           </p>
           {errorMessage && (
             <div className="rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-sm text-rose-100">
@@ -1405,7 +1824,7 @@ function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConf
               disabled={isDeleting}
               className="inline-flex items-center justify-center rounded-full border border-[color:var(--color-border-subtle)] px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] text-[color:var(--color-slate-200)] transition hover:border-[color:var(--color-border-soft)] hover:text-white disabled:cursor-not-allowed disabled:opacity-60"
             >
-              {t('editor.button.cancel')}
+              {t("editor.button.cancel")}
             </button>
             <button
               type="button"
@@ -1413,7 +1832,9 @@ function DeleteTaskModal({ open, onClose, task, isDeleting, errorMessage, onConf
               disabled={isDeleting}
               className="inline-flex items-center justify-center rounded-full bg-rose-600/90 px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] text-white shadow-[0_10px_30px_rgba(225,29,72,0.3)] transition hover:bg-rose-600 disabled:cursor-not-allowed disabled:opacity-60"
             >
-              {isDeleting ? t('editor.modal.delete.loading') : t('editor.modal.delete.confirm')}
+              {isDeleting
+                ? t("editor.modal.delete.loading")
+                : t("editor.modal.delete.confirm")}
             </button>
           </div>
         </div>
@@ -1426,7 +1847,10 @@ function TaskListSkeleton() {
   return (
     <div className="grid gap-3 sm:grid-cols-2 lg:grid-cols-3">
       {Array.from({ length: 6 }).map((_, index) => (
-        <div key={index} className="h-40 animate-pulse rounded-2xl border border-white/5 bg-[color:var(--color-overlay-1)]" />
+        <div
+          key={index}
+          className="h-40 animate-pulse rounded-2xl border border-white/5 bg-[color:var(--color-overlay-1)]"
+        />
       ))}
     </div>
   );
@@ -1443,38 +1867,47 @@ function TaskListEmpty({ message }: { message: string }) {
   );
 }
 
-function TaskListError({ message, onRetry }: { message: string; onRetry: () => void }) {
+function TaskListError({
+  message,
+  onRetry,
+}: {
+  message: string;
+  onRetry: () => void;
+}) {
   const { t } = usePostLoginLanguage();
 
   return (
     <div className="flex flex-col items-center gap-4 rounded-2xl border border-rose-500/30 bg-rose-500/10 px-6 py-8 text-center text-sm text-rose-100">
-      <p className="font-semibold">{t('editor.error.loadTasks.title')}</p>
+      <p className="font-semibold">{t("editor.error.loadTasks.title")}</p>
       <p className="max-w-sm text-rose-200/80">{message}</p>
       <button
         type="button"
         onClick={onRetry}
         className="rounded-full border border-[color:var(--color-border-soft)] bg-[color:var(--color-overlay-2)] px-4 py-2 text-xs font-semibold uppercase tracking-[0.18em] text-white transition hover:border-[color:var(--color-border-strong)]"
       >
-        {t('editor.button.retry')}
+        {t("editor.button.retry")}
       </button>
     </div>
   );
 }
 
-function formatDateLabel(value: string | null, locale: 'es' | 'en' = 'es'): string {
+function formatDateLabel(
+  value: string | null,
+  locale: "es" | "en" = "es",
+): string {
   if (!value) {
-    return '—';
+    return "—";
   }
 
   const parsed = new Date(value);
   if (Number.isNaN(parsed.getTime())) {
-    return '—';
+    return "—";
   }
 
   return parsed.toLocaleDateString(locale);
 }
 
-type ToastMessage = { type: 'success' | 'error' | 'info'; text: string };
+type ToastMessage = { type: "success" | "error" | "info"; text: string };
 
 interface CreateTaskModalProps {
   open: boolean;
@@ -1484,7 +1917,17 @@ interface CreateTaskModalProps {
   isLoadingPillars: boolean;
   pillarsError: Error | null;
   onRetryPillars: () => void;
+  guideStepId?: EditorGuideStepId | null;
 }
+
+type TaskCategorySuggestion = {
+  pillarId: string;
+  pillarLabel: string;
+  traitId: string;
+  traitLabel: string;
+  rationale: string;
+  confidence: number | null;
+};
 
 function CreateTaskModal({
   open,
@@ -1494,13 +1937,30 @@ function CreateTaskModal({
   isLoadingPillars,
   pillarsError,
   onRetryPillars,
+  guideStepId = null,
 }: CreateTaskModalProps) {
   const { language, t } = usePostLoginLanguage();
-  const activeLocale = language === 'es' ? 'es' : 'en';
-  const [selectedPillarId, setSelectedPillarId] = useState('');
-  const [selectedTraitId, setSelectedTraitId] = useState('');
-  const [title, setTitle] = useState('');
-  const [difficultyId, setDifficultyId] = useState('');
+  const activeLocale = language === "es" ? "es" : "en";
+  const [title, setTitle] = useState("");
+  const [difficultyId, setDifficultyId] = useState("");
+  const [suggestionStatus, setSuggestionStatus] = useState<
+    "idle" | "analyzing" | "ready"
+  >("idle");
+  const [flowState, setFlowState] = useState<
+    | "idle"
+    | "analyzing"
+    | "suggestion-ready"
+    | "manual-category-open"
+    | "confirming"
+    | "created"
+    | "error"
+  >("idle");
+  const [suggestion, setSuggestion] = useState<TaskCategorySuggestion | null>(
+    null,
+  );
+  const [manualCategoryEnabled, setManualCategoryEnabled] = useState(false);
+  const [manualPillarId, setManualPillarId] = useState("");
+  const [manualTraitId, setManualTraitId] = useState("");
   const [errors, setErrors] = useState<Record<string, string>>({});
   const [toast, setToast] = useState<ToastMessage | null>(null);
 
@@ -1517,12 +1977,6 @@ function CreateTaskModal({
 
   const { createTask, status: createStatus } = useCreateTask();
   const {
-    data: traits,
-    isLoading: isLoadingTraits,
-    error: traitsError,
-    reload: reloadTraits,
-  } = useTraits(open ? selectedPillarId : null);
-  const {
     data: difficulties,
     isLoading: isLoadingDifficulties,
     error: difficultiesError,
@@ -1535,10 +1989,14 @@ function CreateTaskModal({
 
   useEffect(() => {
     if (!open) {
-      setSelectedPillarId('');
-      setSelectedTraitId('');
-      setTitle('');
-      setDifficultyId('');
+      setTitle("");
+      setDifficultyId("");
+      setSuggestionStatus("idle");
+      setFlowState("idle");
+      setSuggestion(null);
+      setManualCategoryEnabled(false);
+      setManualPillarId("");
+      setManualTraitId("");
       setErrors({});
       setToast(null);
     }
@@ -1546,16 +2004,16 @@ function CreateTaskModal({
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
+      if (event.key === "Escape") {
         event.preventDefault();
         handleClose();
       }
     };
 
     if (open) {
-      window.addEventListener('keydown', handleKeyDown);
+      window.addEventListener("keydown", handleKeyDown);
       return () => {
-        window.removeEventListener('keydown', handleKeyDown);
+        window.removeEventListener("keydown", handleKeyDown);
       };
     }
 
@@ -1570,42 +2028,262 @@ function CreateTaskModal({
     return undefined;
   }, [toast]);
 
-  useEffect(() => {
-    setSelectedTraitId('');
-    clearError('trait');
-  }, [selectedPillarId, clearError]);
-
   const sortedPillars = useMemo(() => {
-    return [...pillars].sort((a, b) => a.name.localeCompare(b.name, activeLocale, { sensitivity: 'base' }));
+    return [...pillars].sort((a, b) =>
+      a.name.localeCompare(b.name, activeLocale, { sensitivity: "base" }),
+    );
   }, [activeLocale, pillars]);
 
-  const filteredTraits = useMemo(() => {
-    return traits.filter((trait) => trait.pillarId === selectedPillarId);
-  }, [traits, selectedPillarId]);
-
   const sortedDifficulties = useMemo(() => {
-    return [...difficulties].sort((a, b) => a.name.localeCompare(b.name, activeLocale, { sensitivity: 'base' }));
+    return [...difficulties].sort((a, b) =>
+      a.name.localeCompare(b.name, activeLocale, { sensitivity: "base" }),
+    );
   }, [activeLocale, difficulties]);
 
-  const isSubmitting = createStatus === 'loading';
+  const selectedManualPillar = useMemo(
+    () => sortedPillars.find((pillar) => pillar.id === manualPillarId) ?? null,
+    [manualPillarId, sortedPillars],
+  );
+
+  const [manualTraits, setManualTraits] = useState<Trait[]>([]);
+  const [isLoadingManualTraits, setIsLoadingManualTraits] = useState(false);
+  const [manualTraitsError, setManualTraitsError] = useState<string | null>(
+    null,
+  );
+
+  useEffect(() => {
+    if (!manualCategoryEnabled || !manualPillarId) {
+      setManualTraits([]);
+      setManualTraitsError(null);
+      setManualTraitId("");
+      return;
+    }
+
+    let cancelled = false;
+    setIsLoadingManualTraits(true);
+    setManualTraitsError(null);
+    void fetchCatalogTraits(manualPillarId)
+      .then((traits) => {
+        if (cancelled) {
+          return;
+        }
+        setManualTraits(traits);
+      })
+      .catch((error) => {
+        if (cancelled) {
+          return;
+        }
+        console.error("Failed to load traits for manual selection", error);
+        setManualTraitsError(t("editor.error.traits.load"));
+        setManualTraits([]);
+      })
+      .finally(() => {
+        if (!cancelled) {
+          setIsLoadingManualTraits(false);
+        }
+      });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [manualCategoryEnabled, manualPillarId, t]);
+
+  const isSubmitting = createStatus === "loading";
+  const isAnalyzing = suggestionStatus === "analyzing";
+  const isGuideAIThinkingStep = guideStepId === "modal-ai-thinking";
+  const [guideSimulationPhase, setGuideSimulationPhase] = useState<
+    "idle" | "analyzing" | "pillar" | "trait"
+  >("idle");
+
+  useEffect(() => {
+    if (isSubmitting) {
+      setFlowState("confirming");
+    }
+  }, [isSubmitting]);
+
+  const mapClassificationToSuggestion = useCallback(
+    (classification: UserTaskClassification): TaskCategorySuggestion | null => {
+      if (!classification.pillarId || !classification.traitId) {
+        return null;
+      }
+
+      const normalizeCode = (code: string | null): string | null => {
+        if (!code) {
+          return null;
+        }
+        return code.trim().toLowerCase();
+      };
+
+      const resolvedPillarId = String(classification.pillarId);
+      const resolvedTraitId = String(classification.traitId);
+
+      const pillarFromCatalog = sortedPillars.find(
+        (pillar) => pillar.id === resolvedPillarId,
+      );
+      const traitFromManualList = manualTraits.find(
+        (trait) => trait.id === resolvedTraitId,
+      );
+
+      return {
+        pillarId: resolvedPillarId,
+        traitId: resolvedTraitId,
+        pillarLabel:
+          classification.pillarName
+          ?? (pillarFromCatalog
+            ? localizePillarLabel(pillarFromCatalog.name, language)
+            : normalizeCode(classification.pillarCode) ?? resolvedPillarId),
+        traitLabel:
+          classification.traitName
+          ?? (traitFromManualList
+            ? localizeTraitLabel(
+              {
+                name: traitFromManualList.name,
+                code: traitFromManualList.code,
+                fallback: traitFromManualList.id,
+              },
+              language,
+            )
+            : normalizeCode(classification.traitCode) ?? resolvedTraitId),
+        rationale:
+          classification.rationale
+          ?? t("editor.modal.aiCreate.suggestedCategory"),
+        confidence: classification.confidence,
+      };
+    },
+    [language, manualTraits, sortedPillars, t],
+  );
+
+  useEffect(() => {
+    if (!isGuideAIThinkingStep) {
+      setGuideSimulationPhase("idle");
+      return;
+    }
+
+    setGuideSimulationPhase("analyzing");
+    const toPillar = window.setTimeout(
+      () => setGuideSimulationPhase("pillar"),
+      720,
+    );
+    const toTrait = window.setTimeout(
+      () => setGuideSimulationPhase("trait"),
+      1320,
+    );
+
+    return () => {
+      window.clearTimeout(toPillar);
+      window.clearTimeout(toTrait);
+    };
+  }, [isGuideAIThinkingStep]);
+
+  const guideSuggestion = isGuideAIThinkingStep
+    ? {
+        pillarId: "guide-soul",
+        pillarLabel: language === "es" ? "Alma" : "Soul",
+        traitId: "guide-trait-gratitude",
+        traitLabel: language === "es" ? "Gratitud" : "Gratitude",
+        rationale:
+          language === "es"
+            ? "La simulación muestra una clasificación coherente: Alma + Gratitud."
+            : "The simulation shows a coherent classification: Soul + Gratitude.",
+        confidence: null,
+      }
+    : null;
+  const visibleSuggestion = suggestion ?? guideSuggestion;
+  const shouldShowPillarChip =
+    !isGuideAIThinkingStep || guideSimulationPhase !== "analyzing";
+  const shouldShowTraitChip =
+    !isGuideAIThinkingStep || guideSimulationPhase === "trait";
+  const showAnalyzingCard = isAnalyzing || isGuideAIThinkingStep;
+  const isManualCategoryOpen = flowState === "manual-category-open";
+  const hasManualCategorySelection = Boolean(manualPillarId && manualTraitId);
+  const shouldUseManualCategory =
+    manualCategoryEnabled && hasManualCategorySelection;
   const isSubmitDisabled =
-    isSubmitting || !selectedPillarId || !selectedTraitId || title.trim().length === 0 || !userId;
+    isSubmitting ||
+    (!suggestion && !hasManualCategorySelection) ||
+    title.trim().length === 0 ||
+    !userId;
+  const isSuggestDisabled =
+    (isAnalyzing && !isGuideAIThinkingStep) ||
+    title.trim().length === 0 ||
+    !userId;
+
+  const handleSuggestCategory = useCallback(async () => {
+    const validationErrors: Record<string, string> = {};
+    if (title.trim().length === 0) {
+      validationErrors.title = t("editor.validation.titleRequired");
+    }
+    if (!userId) {
+      validationErrors.user = t("editor.validation.userNotFound");
+    }
+    setErrors(validationErrors);
+    if (Object.keys(validationErrors).length > 0) {
+      return;
+    }
+
+    setSuggestionStatus("analyzing");
+    setFlowState("analyzing");
+    setSuggestion(null);
+    clearError("suggestion");
+
+    try {
+      const classification = await classifyUserTask(userId!, {
+        title: title.trim(),
+      });
+
+      const resolvedSuggestion = mapClassificationToSuggestion(classification);
+      if (!resolvedSuggestion || classification.requiresManualSelection) {
+        setSuggestionStatus("idle");
+        setFlowState("manual-category-open");
+        setManualCategoryEnabled(true);
+        setErrors((previous) => ({
+          ...previous,
+          suggestion: t("editor.modal.aiCreate.catalogFallback"),
+        }));
+        return;
+      }
+
+      setSuggestion(resolvedSuggestion);
+      setSuggestionStatus("ready");
+      setFlowState("suggestion-ready");
+      if (manualCategoryEnabled) {
+        setManualCategoryEnabled(false);
+      }
+    } catch (error) {
+      console.error("Failed to resolve AI suggestion for lab editor", error);
+      setSuggestionStatus("idle");
+      setFlowState("error");
+      setManualCategoryEnabled(true);
+      setErrors((previous) => ({
+        ...previous,
+        suggestion: t("editor.modal.aiCreate.suggestionError"),
+      }));
+    }
+  }, [
+    clearError,
+    manualCategoryEnabled,
+    mapClassificationToSuggestion,
+    t,
+    title,
+    userId,
+  ]);
 
   const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
     const validationErrors: Record<string, string> = {};
-    if (!selectedPillarId) {
-      validationErrors.pillar = t('editor.validation.selectPillar');
-    }
-    if (!selectedTraitId) {
-      validationErrors.trait = t('editor.validation.selectTrait');
-    }
     if (title.trim().length === 0) {
-      validationErrors.title = t('editor.validation.titleRequired');
+      validationErrors.title = t("editor.validation.titleRequired");
+    }
+    if (!suggestion) {
+      if (!hasManualCategorySelection) {
+        validationErrors.suggestion = t(
+          "editor.modal.aiCreate.confirmationRequired",
+        );
+      }
     }
     if (!userId) {
-      validationErrors.user = t('editor.validation.userNotFound');
+      validationErrors.user = t("editor.validation.userNotFound");
     }
 
     setErrors(validationErrors);
@@ -1615,20 +2293,34 @@ function CreateTaskModal({
     }
 
     try {
+      const resolvedPillarId = shouldUseManualCategory
+        ? manualPillarId
+        : suggestion?.pillarId ?? manualPillarId;
+      const resolvedTraitId = shouldUseManualCategory
+        ? manualTraitId
+        : suggestion?.traitId ?? manualTraitId;
       await createTask(userId!, {
         title: title.trim(),
-        pillarId: selectedPillarId,
-        traitId: selectedTraitId,
+        pillarId: resolvedPillarId,
+        traitId: resolvedTraitId,
         statId: null,
         difficultyId: difficultyId || null,
       });
-      setToast({ type: 'success', text: t('editor.toast.create.success') });
-      setTitle('');
-      setDifficultyId('');
+      setToast({ type: "success", text: t("editor.toast.create.success") });
+      setFlowState("created");
+      setTitle("");
+      setDifficultyId("");
+      setSuggestionStatus("idle");
+      setSuggestion(null);
+      setManualCategoryEnabled(false);
+      setManualPillarId("");
+      setManualTraitId("");
       setErrors({});
     } catch (error) {
-      const message = error instanceof Error ? error.message : t('editor.toast.create.error');
-      setToast({ type: 'error', text: message });
+      const message =
+        error instanceof Error ? error.message : t("editor.toast.create.error");
+      setFlowState("error");
+      setToast({ type: "error", text: message });
     }
   };
 
@@ -1637,176 +2329,331 @@ function CreateTaskModal({
   }
 
   return (
-    <div className="create-task-modal__overlay fixed inset-0 z-[60] flex items-end justify-center bg-slate-950/70 backdrop-blur-sm md:items-center" data-light-scope="editor">
+    <div
+      className="create-task-modal__overlay fixed inset-0 z-[60] flex items-end justify-center bg-slate-950/70 backdrop-blur-sm md:items-center"
+      data-light-scope="editor"
+    >
       <button
         type="button"
-        aria-label={t('editor.button.close')}
+        aria-label={t("editor.button.close")}
         onClick={handleClose}
         className="absolute inset-0 h-full w-full"
       />
       <div className="relative z-10 w-full max-w-2xl p-4">
         <form
           onSubmit={handleSubmit}
+          data-editor-guide-target="new-task-modal-dialog"
           className="create-task-modal__dialog max-h-[90vh] overflow-y-auto rounded-2xl border p-6"
           onClick={(event) => event.stopPropagation()}
         >
           <div className="create-task-modal space-y-6">
             <header className="space-y-1">
-              <p className="create-task-modal__badge text-[11px] font-semibold uppercase tracking-[0.24em]">{t('editor.modal.create.badge')}</p>
-              <h2 className="create-task-modal__title text-xl font-semibold">{t('editor.modal.create.title')}</h2>
-              <p className="create-task-modal__description text-sm">
-                {t('editor.modal.create.description')}
-              </p>
+              <div className="flex items-start justify-between gap-4">
+                <div className="space-y-1">
+                  <p className="create-task-ai-modal__badge text-[11px] font-semibold uppercase tracking-[0.24em]">
+                    {t("editor.modal.aiCreate.badge")}
+                  </p>
+                  <h2 className="create-task-ai-modal__title text-xl font-semibold">
+                    {t("editor.modal.aiCreate.title")}
+                  </h2>
+                  <p className="create-task-ai-modal__description text-sm">
+                    {t("editor.modal.aiCreate.description")}
+                  </p>
+                </div>
+                <button
+                  type="button"
+                  aria-label={t("editor.button.close")}
+                  className="create-task-ai-modal__close inline-flex h-9 w-9 items-center justify-center rounded-full border text-lg transition"
+                  onClick={handleClose}
+                >
+                  ×
+                </button>
+              </div>
             </header>
 
-            <section className="space-y-4">
-              <div className="space-y-1.5">
-                <p className="create-task-modal__section-label text-[11px] font-semibold uppercase tracking-[0.24em]">{t('editor.modal.create.step1')}</p>
-                <div className="flex flex-col gap-2">
-                  <select
-                    value={selectedPillarId}
-                    onChange={(event) => {
-                      setSelectedPillarId(event.target.value);
-                      clearError('pillar');
-                    }}
-                    className="create-task-modal__control create-task-modal__control--pill w-full appearance-none rounded-full border px-4 py-2.5 text-sm ios-touch-input focus:outline-none disabled:cursor-not-allowed"
-                    disabled={isLoadingPillars || pillarsError != null}
-                  >
-                    <option value="" className="create-task-modal__option">
-                      {t('editor.modal.create.selectPillarPlaceholder')}
-                    </option>
-                    {sortedPillars.map((pillar) => (
-                      <option key={pillar.id} value={pillar.id} className="create-task-modal__option">
-                        {localizePillarLabel(pillar.name, language)}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                {isLoadingPillars && (
-                  <p className="create-task-modal__hint text-[11px] uppercase tracking-[0.2em]">{t('editor.loading.pillars')}</p>
-                )}
-                {pillarsError && (
-                  <div className="space-y-1 rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-xs text-rose-100">
-                    <p>{t('editor.error.pillars.load')}</p>
-                    <button
-                      type="button"
-                      onClick={onRetryPillars}
-                      className="font-semibold text-rose-200 underline decoration-dotted"
-                    >
-                      {t('editor.button.retry')}
-                    </button>
-                  </div>
-                )}
-                {errors.pillar && <p className="text-xs text-rose-300">{errors.pillar}</p>}
-                {!isLoadingPillars && !pillarsError && sortedPillars.length === 0 && (
-                  <p className="create-task-modal__hint text-xs">No encontramos pilares disponibles por ahora.</p>
-                )}
-              </div>
-
-              <div className="space-y-1.5">
-                <p className="create-task-modal__section-label text-[11px] font-semibold uppercase tracking-[0.24em]">{t('editor.modal.create.step2')}</p>
-                <div className="flex flex-col gap-2">
-                  <select
-                    value={selectedTraitId}
-                    onChange={(event) => {
-                      setSelectedTraitId(event.target.value);
-                      clearError('trait');
-                    }}
-                    className="create-task-modal__control create-task-modal__control--pill w-full appearance-none rounded-full border px-4 py-2.5 text-sm ios-touch-input focus:outline-none disabled:cursor-not-allowed"
-                    disabled={!selectedPillarId || isLoadingTraits}
-                  >
-                    <option value="" className="create-task-modal__option">
-                      {selectedPillarId ? t('editor.modal.create.selectTraitPlaceholder') : t('editor.modal.create.selectPillarFirst')}
-                    </option>
-                    {filteredTraits.map((trait) => (
-                      <option key={trait.id} value={trait.id} className="create-task-modal__option">
-                        {localizeTraitLabel({ name: trait.name, code: trait.code, fallback: trait.id }, language)}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                {isLoadingTraits && (
-                  <p className="create-task-modal__hint text-[11px] uppercase tracking-[0.2em]">{t('editor.loading.traits')}</p>
-                )}
-                {traitsError && (
-                  <div className="space-y-1 rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-xs text-rose-100">
-                    <p>{t('editor.error.traits.load')}</p>
-                    <button
-                      type="button"
-                      onClick={reloadTraits}
-                      className="font-semibold text-rose-200 underline decoration-dotted"
-                    >
-                      {t('editor.button.retry')}
-                    </button>
-                  </div>
-                )}
-                {errors.trait && <p className="text-xs text-rose-300">{errors.trait}</p>}
-                {selectedPillarId && !isLoadingTraits && filteredTraits.length === 0 && !traitsError && (
-                  <p className="create-task-modal__hint text-xs">{t('editor.empty.noTraits')}</p>
-                )}
-              </div>
-
-            </section>
-
-            <section className="space-y-4">
+            <section
+              className="space-y-4"
+              data-editor-guide-target="new-task-modal-core"
+            >
               <div className="space-y-2">
                 <label className="flex flex-col gap-2">
-                  <span className="create-task-modal__field-label text-xs font-semibold uppercase tracking-[0.18em]">{t('editor.modal.create.taskTitleLabel')}</span>
-                  <input
-                    type="text"
+                  <span className="create-task-ai-modal__field-label text-xs font-semibold uppercase tracking-[0.18em]">
+                    {t("editor.modal.aiCreate.taskTitleLabel")}
+                  </span>
+                  <textarea
+                    data-editor-guide-target="new-task-modal-input"
                     value={title}
                     onChange={(event) => {
                       setTitle(event.target.value);
-                      clearError('title');
+                      clearError("title");
+                      clearError("suggestion");
+                      setSuggestionStatus("idle");
+                      setFlowState("idle");
+                      setSuggestion(null);
                     }}
-                    placeholder={t('editor.modal.taskTitle.placeholder')}
-                    className="create-task-modal__control w-full rounded-2xl border px-4 py-3 text-sm ios-touch-input focus:outline-none"
+                    placeholder={t("editor.modal.aiCreate.taskTitlePlaceholder")}
+                    className="create-task-ai-modal__control w-full rounded-2xl border px-4 py-3 text-sm ios-touch-input focus:outline-none"
+                    rows={3}
                   />
                 </label>
-                {errors.title && <p className="text-xs text-rose-300">{errors.title}</p>}
+                {errors.title && (
+                  <p className="text-xs text-rose-300">{errors.title}</p>
+                )}
               </div>
 
               <div className="space-y-2">
                 <label className="flex flex-col gap-2">
-                  <span className="create-task-modal__field-label text-xs font-semibold uppercase tracking-[0.18em]">{t('editor.field.difficulty')}</span>
+                  <span className="create-task-ai-modal__field-label text-xs font-semibold uppercase tracking-[0.18em]">
+                    {t("editor.field.difficulty")}
+                  </span>
                   <select
+                    data-editor-guide-target="new-task-modal-difficulty"
                     value={difficultyId}
                     onChange={(event) => setDifficultyId(event.target.value)}
-                    className="create-task-modal__control w-full appearance-none rounded-2xl border px-4 py-3 text-sm ios-touch-input focus:outline-none disabled:cursor-not-allowed"
+                    className="create-task-ai-modal__control w-full appearance-none rounded-2xl border px-4 py-3 text-sm ios-touch-input focus:outline-none disabled:cursor-not-allowed"
                     disabled={isLoadingDifficulties}
                   >
-                    <option value="" className="create-task-modal__option">
-                      {t('editor.modal.create.selectDifficultyPlaceholder')}
+                    <option value="" className="create-task-ai-modal__option">
+                      {t("editor.modal.create.selectDifficultyPlaceholder")}
                     </option>
                     {sortedDifficulties.map((difficulty) => (
-                      <option key={difficulty.id} value={difficulty.id} className="create-task-modal__option">
+                      <option
+                        key={difficulty.id}
+                        value={difficulty.id}
+                        className="create-task-ai-modal__option"
+                      >
                         {localizeDifficultyLabel(difficulty.name, language)}
                       </option>
                     ))}
                   </select>
                 </label>
                 {isLoadingDifficulties && (
-                  <p className="create-task-modal__hint text-[11px] uppercase tracking-[0.2em]">{t('editor.loading.difficulties')}</p>
+                  <p className="create-task-modal__hint text-[11px] uppercase tracking-[0.2em]">
+                    {t("editor.loading.difficulties")}
+                  </p>
                 )}
                 {difficultiesError && (
                   <div className="space-y-1 rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-xs text-rose-100">
-                    <p>{t('editor.error.difficulties.load')}</p>
+                    <p>{t("editor.error.difficulties.load")}</p>
                     <button
                       type="button"
                       onClick={reloadDifficulties}
                       className="font-semibold text-rose-200 underline decoration-dotted"
                     >
-                      {t('editor.button.retry')}
+                      {t("editor.button.retry")}
                     </button>
                   </div>
                 )}
               </div>
-
             </section>
 
-            {errors.user && <p className="text-xs text-rose-300">{errors.user}</p>}
+            <section
+              className="space-y-3"
+              data-editor-guide-target="new-task-modal-ai-zone"
+            >
+              <button
+                type="button"
+                data-editor-guide-target="new-task-modal-ai-action"
+                className="create-task-ai-modal__suggest-button inline-flex w-full items-center justify-center gap-2 rounded-xl px-4 py-2.5 text-xs font-semibold transition disabled:cursor-not-allowed disabled:opacity-60 sm:w-auto"
+                onClick={() => void handleSuggestCategory()}
+                disabled={isSuggestDisabled}
+              >
+                <span aria-hidden>✨</span>
+                {isAnalyzing
+                  ? t("editor.modal.aiCreate.analyzing")
+                  : t("editor.modal.aiCreate.suggestButton")}
+              </button>
+              {isLoadingPillars && (
+                <p className="create-task-ai-modal__hint text-[11px] uppercase tracking-[0.2em]">
+                  {t("editor.loading.pillars")}
+                </p>
+              )}
+              {pillarsError && (
+                <div className="space-y-1 rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-xs text-rose-100">
+                  <p>{t("editor.error.pillars.load")}</p>
+                  <button
+                    type="button"
+                    onClick={onRetryPillars}
+                    className="font-semibold text-rose-200 underline decoration-dotted"
+                  >
+                    {t("editor.button.retry")}
+                  </button>
+                </div>
+              )}
+              {showAnalyzingCard && (
+                <section className="create-task-ai-modal__analysis-card space-y-2 rounded-xl border p-3">
+                  <div className="create-task-ai-modal__pulse h-2 w-24 rounded-full" />
+                  <p className="text-sm font-semibold">
+                    {t("editor.modal.aiCreate.analyzing")}
+                  </p>
+                  <p className="create-task-ai-modal__hint text-xs">
+                    {t("editor.modal.aiCreate.analyzingHint")}
+                  </p>
+                </section>
+              )}
 
-            {toast && <ToastBanner tone={toast.type} message={toast.text} className="px-3" />}
+              {visibleSuggestion &&
+                (suggestionStatus === "ready" || isGuideAIThinkingStep) && (
+                  <section
+                    className="create-task-ai-modal__suggestion-strip space-y-3.5 py-2"
+                    data-editor-guide-target="new-task-modal-ai-result"
+                  >
+                    <p className="create-task-ai-modal__field-label text-center text-[11px] font-semibold uppercase tracking-[0.24em]">
+                      {t("editor.modal.aiCreate.suggestedCategory")}
+                    </p>
+                    <div className="flex items-center justify-center gap-2 text-base">
+                      {shouldShowPillarChip ? (
+                        <span className="create-task-ai-modal__result-pill rounded-full border px-4 py-1.5 font-semibold">
+                          {visibleSuggestion.pillarLabel}
+                        </span>
+                      ) : (
+                        <span className="create-task-ai-modal__hint text-xs">
+                          {language === "es"
+                            ? "Detectando pilar…"
+                            : "Detecting pillar…"}
+                        </span>
+                      )}
+                      <span className="create-task-ai-modal__hint">/</span>
+                      {shouldShowTraitChip ? (
+                        <span className="create-task-ai-modal__result-pill rounded-full border px-4 py-1.5 font-semibold">
+                          {visibleSuggestion.traitLabel}
+                        </span>
+                      ) : (
+                        <span className="create-task-ai-modal__hint text-xs">
+                          {language === "es"
+                            ? "Detectando rasgo…"
+                            : "Detecting trait…"}
+                        </span>
+                      )}
+                    </div>
+                    <p className="create-task-ai-modal__hint text-center text-sm">
+                      {visibleSuggestion.rationale}
+                    </p>
+                    <div className="flex flex-wrap items-center justify-center gap-4 pt-1">
+                      <button
+                        type="button"
+                        className="create-task-ai-modal__retry text-xs font-semibold underline decoration-dotted underline-offset-4"
+                        onClick={() => void handleSuggestCategory()}
+                      >
+                        {t("editor.modal.aiCreate.retrySuggestion")}
+                      </button>
+                      <button
+                        type="button"
+                        className="create-task-ai-modal__retry text-xs font-semibold underline decoration-dotted underline-offset-4"
+                        onClick={() => {
+                          setManualCategoryEnabled((previous) => {
+                            if (previous) {
+                              setManualPillarId("");
+                              setManualTraitId("");
+                              setFlowState("suggestion-ready");
+                            } else {
+                              setFlowState("manual-category-open");
+                            }
+                            return !previous;
+                          });
+                          clearError("suggestion");
+                        }}
+                      >
+                        {manualCategoryEnabled
+                          ? t("editor.modal.aiCreate.useAiSuggestion")
+                          : t("editor.modal.aiCreate.manualCategory")}
+                      </button>
+                    </div>
+                  </section>
+                )}
+
+              {(manualCategoryEnabled || isManualCategoryOpen) && (
+                <section
+                  className="create-task-ai-modal__manual-grid grid gap-3 rounded-xl border p-3"
+                  data-editor-guide-target="new-task-modal-ai-result"
+                >
+                  <div className="flex justify-end">
+                    <button
+                      type="button"
+                      className="create-task-ai-modal__retry text-xs font-semibold underline decoration-dotted underline-offset-4"
+                      onClick={() => {
+                        setManualCategoryEnabled(false);
+                        setManualPillarId("");
+                        setManualTraitId("");
+                        setFlowState("suggestion-ready");
+                        clearError("suggestion");
+                      }}
+                    >
+                      {t("editor.modal.aiCreate.useAiSuggestion")}
+                    </button>
+                  </div>
+                  <label className="flex flex-col gap-2">
+                    <span className="create-task-ai-modal__field-label text-[11px] font-semibold uppercase tracking-[0.2em]">
+                      {t("editor.field.pillar")}
+                    </span>
+                    <select
+                      value={manualPillarId}
+                      onChange={(event) => {
+                        setManualPillarId(event.target.value);
+                        setManualTraitId("");
+                        clearError("suggestion");
+                      }}
+                      className="create-task-ai-modal__control w-full appearance-none rounded-xl border px-3 py-2 text-sm ios-touch-input focus:outline-none"
+                    >
+                      <option value="">
+                        {t("editor.modal.create.selectPillarPlaceholder")}
+                      </option>
+                      {sortedPillars.map((pillar) => (
+                        <option key={pillar.id} value={pillar.id}>
+                          {localizePillarLabel(pillar.name, language)}
+                        </option>
+                      ))}
+                    </select>
+                  </label>
+                  <label className="flex flex-col gap-2">
+                    <span className="create-task-ai-modal__field-label text-[11px] font-semibold uppercase tracking-[0.2em]">
+                      {t("editor.field.trait")}
+                    </span>
+                    <select
+                      value={manualTraitId}
+                      onChange={(event) => {
+                        setManualTraitId(event.target.value);
+                        clearError("suggestion");
+                      }}
+                      className="create-task-ai-modal__control w-full appearance-none rounded-xl border px-3 py-2 text-sm ios-touch-input focus:outline-none"
+                      disabled={!selectedManualPillar || isLoadingManualTraits}
+                    >
+                      <option value="">
+                        {selectedManualPillar
+                          ? t("editor.modal.create.selectTraitPlaceholder")
+                          : t("editor.modal.create.selectPillarFirst")}
+                      </option>
+                      {manualTraits.map((trait) => (
+                        <option key={trait.id} value={trait.id}>
+                          {localizeTraitLabel(
+                            { name: trait.name, code: trait.code, fallback: trait.id },
+                            language,
+                          )}
+                        </option>
+                      ))}
+                    </select>
+                  </label>
+                  {manualTraitsError && (
+                    <p className="text-xs text-rose-300">{manualTraitsError}</p>
+                  )}
+                </section>
+              )}
+            </section>
+
+            {errors.suggestion && (
+              <p className="text-xs text-rose-300">{errors.suggestion}</p>
+            )}
+            {errors.user && (
+              <p className="text-xs text-rose-300">{errors.user}</p>
+            )}
+
+            {toast && (
+              <ToastBanner
+                tone={toast.type}
+                message={toast.text}
+                className="px-3"
+              />
+            )}
 
             <div className="flex flex-col gap-3 pt-2 sm:flex-row sm:justify-end">
               <button
@@ -1814,15 +2661,20 @@ function CreateTaskModal({
                 onClick={handleClose}
                 className="create-task-modal__button-secondary inline-flex items-center justify-center rounded-full border px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] transition"
               >
-                {t('editor.button.cancel')}
+                {t("editor.button.cancel")}
               </button>
               <button
                 type="submit"
                 disabled={isSubmitDisabled}
                 className="inline-flex items-center justify-center rounded-full px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] text-[#121212] transition disabled:cursor-not-allowed disabled:opacity-60"
-                style={{ background: 'var(--gradient-innerbloom)', boxShadow: 'var(--shadow-innerbloom-cta)' }}
+                style={{
+                  background: "var(--gradient-innerbloom)",
+                  boxShadow: "var(--shadow-innerbloom-cta)",
+                }}
               >
-                {isSubmitting ? t('editor.button.creating') : t('editor.button.createTask')}
+                {isSubmitting
+                  ? t("editor.button.creating")
+                  : t("editor.modal.aiCreate.confirmButton")}
               </button>
             </div>
           </div>
@@ -1839,7 +2691,7 @@ interface EditTaskModalProps {
   userId: string | null;
   task: UserTask | null;
   pillars: Pillar[];
-  variant?: 'modal' | 'panel';
+  variant?: "modal" | "panel";
   navigationTasks?: UserTask[];
   onNavigateTask?: (taskId: string) => void;
 }
@@ -1851,14 +2703,14 @@ function EditTaskModal({
   userId,
   task,
   pillars,
-  variant = 'modal',
+  variant = "modal",
   navigationTasks = [],
   onNavigateTask,
 }: EditTaskModalProps) {
   const { language, t } = usePostLoginLanguage();
-  const activeLocale = language === 'es' ? 'es' : 'en';
-  const [title, setTitle] = useState('');
-  const [difficultyId, setDifficultyId] = useState('');
+  const activeLocale = language === "es" ? "es" : "en";
+  const [title, setTitle] = useState("");
+  const [difficultyId, setDifficultyId] = useState("");
   const [isActive, setIsActive] = useState(true);
   const [errors, setErrors] = useState<Record<string, string>>({});
   const [toast, setToast] = useState<ToastMessage | null>(null);
@@ -1875,27 +2727,35 @@ function EditTaskModal({
   }, []);
 
   const { updateTask, status: updateStatus } = useUpdateTask();
-  const { data: difficulties, isLoading: isLoadingDifficulties, error: difficultiesError, reload: reloadDifficulties } =
-    useDifficulties();
+  const {
+    data: difficulties,
+    isLoading: isLoadingDifficulties,
+    error: difficultiesError,
+    reload: reloadDifficulties,
+  } = useDifficulties();
 
   const currentPillarId = open && task?.pillarId ? task.pillarId : null;
   const { data: traits } = useTraits(currentPillarId);
 
   const sortedDifficulties = useMemo(() => {
-    return [...difficulties].sort((a, b) => a.name.localeCompare(b.name, activeLocale, { sensitivity: 'base' }));
+    return [...difficulties].sort((a, b) =>
+      a.name.localeCompare(b.name, activeLocale, { sensitivity: "base" }),
+    );
   }, [activeLocale, difficulties]);
 
   const pillarName = useMemo(() => {
     if (!task?.pillarId) {
-      return t('editor.symbol.empty');
+      return t("editor.symbol.empty");
     }
-    const sourceLabel = pillars.find((pillar) => pillar.id === task.pillarId)?.name ?? task.pillarId;
+    const sourceLabel =
+      pillars.find((pillar) => pillar.id === task.pillarId)?.name ??
+      task.pillarId;
     return localizePillarLabel(sourceLabel, language);
   }, [language, pillars, task?.pillarId, t]);
 
   const traitName = useMemo(() => {
     if (!task?.traitId) {
-      return t('editor.symbol.empty');
+      return t("editor.symbol.empty");
     }
     const trait = traits.find((entry) => entry.id === task.traitId);
     return localizeTraitLabel(
@@ -1904,7 +2764,7 @@ function EditTaskModal({
     );
   }, [language, t, task?.traitId, traits]);
 
-  const isSubmitting = updateStatus === 'loading';
+  const isSubmitting = updateStatus === "loading";
 
   const handleClose = useCallback(() => {
     onClose();
@@ -1912,8 +2772,8 @@ function EditTaskModal({
 
   useEffect(() => {
     if (!open) {
-      setTitle('');
-      setDifficultyId('');
+      setTitle("");
+      setDifficultyId("");
       setIsActive(true);
       setErrors({});
       setToast(null);
@@ -1921,8 +2781,8 @@ function EditTaskModal({
     }
 
     if (task) {
-      setTitle(task.title ?? '');
-      setDifficultyId(task.difficultyId ?? '');
+      setTitle(task.title ?? "");
+      setDifficultyId(task.difficultyId ?? "");
       setIsActive(Boolean(task.isActive));
       setErrors({});
       setToast(null);
@@ -1931,16 +2791,16 @@ function EditTaskModal({
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
+      if (event.key === "Escape") {
         event.preventDefault();
         handleClose();
       }
     };
 
     if (open) {
-      window.addEventListener('keydown', handleKeyDown);
+      window.addEventListener("keydown", handleKeyDown);
       return () => {
-        window.removeEventListener('keydown', handleKeyDown);
+        window.removeEventListener("keydown", handleKeyDown);
       };
     }
 
@@ -1961,15 +2821,15 @@ function EditTaskModal({
     const validationErrors: Record<string, string> = {};
 
     if (!title.trim()) {
-      validationErrors.title = t('editor.validation.titleRequired');
+      validationErrors.title = t("editor.validation.titleRequired");
     }
 
     if (!userId) {
-      validationErrors.user = t('editor.validation.userNotFound');
+      validationErrors.user = t("editor.validation.userNotFound");
     }
 
     if (!task) {
-      validationErrors.task = t('editor.validation.taskNotFoundEdit');
+      validationErrors.task = t("editor.validation.taskNotFoundEdit");
     }
 
     setErrors(validationErrors);
@@ -1990,10 +2850,11 @@ function EditTaskModal({
         return;
       }
 
-      setToast({ type: 'success', text: t('editor.toast.update.success') });
+      setToast({ type: "success", text: t("editor.toast.update.success") });
     } catch (error) {
-      const message = error instanceof Error ? error.message : t('editor.toast.update.error');
-      setToast({ type: 'error', text: message });
+      const message =
+        error instanceof Error ? error.message : t("editor.toast.update.error");
+      setToast({ type: "error", text: message });
     }
   };
 
@@ -2010,103 +2871,133 @@ function EditTaskModal({
     return null;
   }
 
-  const showNavigation = variant === 'panel' && navigationTasks.length > 0 && task != null;
-  const activeIndex = showNavigation ? navigationTasks.findIndex((entry) => entry.id === task.id) : -1;
-  const previousTask = showNavigation && activeIndex > 0 ? navigationTasks[activeIndex - 1] : null;
+  const showNavigation =
+    variant === "panel" && navigationTasks.length > 0 && task != null;
+  const activeIndex = showNavigation
+    ? navigationTasks.findIndex((entry) => entry.id === task.id)
+    : -1;
+  const previousTask =
+    showNavigation && activeIndex > 0 ? navigationTasks[activeIndex - 1] : null;
   const nextTask =
-    showNavigation && activeIndex >= 0 && activeIndex < navigationTasks.length - 1
+    showNavigation &&
+    activeIndex >= 0 &&
+    activeIndex < navigationTasks.length - 1
       ? navigationTasks[activeIndex + 1]
       : null;
   const navigationLabel = showNavigation
     ? `${activeIndex + 1}/${navigationTasks.length}`
     : navigationTasks.length > 0
       ? `—/${navigationTasks.length}`
-      : '—';
+      : "—";
 
   const formBody = (
     <div className="edit-task-modal space-y-6">
       <header className="space-y-1">
-        <p className="text-[11px] font-semibold uppercase tracking-[0.24em] text-[color:var(--color-slate-400)]">{t('editor.modal.edit.badge')}</p>
-        <h2 className="edit-task-modal__title text-xl font-semibold">{t('editor.modal.edit.title')}</h2>
+        <p className="text-[11px] font-semibold uppercase tracking-[0.24em] text-[color:var(--color-slate-400)]">
+          {t("editor.modal.edit.badge")}
+        </p>
+        <h2 className="edit-task-modal__title text-xl font-semibold">
+          {t("editor.modal.edit.title")}
+        </h2>
         <p className="edit-task-modal__description text-sm">
-          {t('editor.modal.edit.description')}
+          {t("editor.modal.edit.description")}
         </p>
       </header>
 
       <section className="space-y-4">
         <div className="space-y-2">
           <span className="edit-task-modal__locked-section-label text-[11px] font-semibold uppercase tracking-[0.24em] text-slate-500">
-            {t('editor.modal.edit.context')}
+            {t("editor.modal.edit.context")}
           </span>
           <div className="grid gap-3 md:grid-cols-2">
-            <ReadOnlyField label={t('editor.field.pillar')} value={pillarName} />
-            <ReadOnlyField label={t('editor.field.trait')} value={traitName} />
+            <ReadOnlyField
+              label={t("editor.field.pillar")}
+              value={pillarName}
+            />
+            <ReadOnlyField label={t("editor.field.trait")} value={traitName} />
           </div>
           <p className="edit-task-modal__locked-section-label text-[11px] uppercase tracking-[0.2em] text-slate-500">
-            {t('editor.modal.edit.lockedHint')}
+            {t("editor.modal.edit.lockedHint")}
           </p>
         </div>
       </section>
 
       <section className="space-y-4">
         <p className="edit-task-modal__editable-section-label text-[11px] font-semibold uppercase tracking-[0.2em]">
-          {t('editor.modal.edit.editableFields')}
+          {t("editor.modal.edit.editableFields")}
         </p>
         <div className="space-y-2">
           <label className="flex flex-col gap-2">
-            <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">{t('editor.modal.create.taskTitleLabel')}</span>
+            <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">
+              {t("editor.modal.create.taskTitleLabel")}
+            </span>
             <input
               type="text"
               value={title}
               onChange={(event) => {
                 setTitle(event.target.value);
-                clearError('title');
+                clearError("title");
               }}
-              placeholder={t('editor.modal.taskTitle.placeholder')}
+              placeholder={t("editor.modal.taskTitle.placeholder")}
               className="edit-task-modal__editable-control w-full rounded-2xl border px-4 py-3 text-sm ios-touch-input transition focus:outline-none"
             />
           </label>
-          {errors.title && <p className="text-xs text-rose-300">{errors.title}</p>}
+          {errors.title && (
+            <p className="text-xs text-rose-300">{errors.title}</p>
+          )}
         </div>
 
         <div className="space-y-2">
           <label className="flex flex-col gap-2">
-            <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">{t('editor.field.difficulty')}</span>
+            <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">
+              {t("editor.field.difficulty")}
+            </span>
             <select
               value={difficultyId}
               onChange={(event) => setDifficultyId(event.target.value)}
               className="edit-task-modal__editable-control w-full appearance-none rounded-2xl border px-4 py-3 text-sm ios-touch-input transition focus:outline-none disabled:cursor-not-allowed"
               disabled={isLoadingDifficulties}
             >
-              <option value="" className="bg-slate-900 text-[color:var(--color-slate-100)]">
-                {t('editor.modal.edit.noDifficultyAssigned')}
+              <option
+                value=""
+                className="bg-slate-900 text-[color:var(--color-slate-100)]"
+              >
+                {t("editor.modal.edit.noDifficultyAssigned")}
               </option>
               {sortedDifficulties.map((difficulty) => (
-                <option key={difficulty.id} value={difficulty.id} className="bg-slate-900 text-[color:var(--color-slate-100)]">
+                <option
+                  key={difficulty.id}
+                  value={difficulty.id}
+                  className="bg-slate-900 text-[color:var(--color-slate-100)]"
+                >
                   {localizeDifficultyLabel(difficulty.name, language)}
                 </option>
               ))}
             </select>
           </label>
           {isLoadingDifficulties && (
-            <p className="text-[11px] uppercase tracking-[0.2em] text-slate-500">{t('editor.loading.difficulties')}</p>
+            <p className="text-[11px] uppercase tracking-[0.2em] text-slate-500">
+              {t("editor.loading.difficulties")}
+            </p>
           )}
           {difficultiesError && (
             <div className="space-y-1 rounded-xl border border-rose-500/40 bg-rose-500/10 px-3 py-2 text-xs text-rose-100">
-              <p>{t('editor.error.difficulties.load')}</p>
+              <p>{t("editor.error.difficulties.load")}</p>
               <button
                 type="button"
                 onClick={reloadDifficulties}
                 className="font-semibold text-rose-200 underline decoration-dotted"
               >
-                {t('editor.button.retry')}
+                {t("editor.button.retry")}
               </button>
             </div>
           )}
         </div>
 
         <div className="space-y-2">
-          <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">{t('editor.field.status')}</span>
+          <span className="edit-task-modal__editable-field-label text-xs font-semibold uppercase tracking-[0.18em]">
+            {t("editor.field.status")}
+          </span>
           <label className="flex items-center gap-3">
             <input
               type="checkbox"
@@ -2114,7 +3005,11 @@ function EditTaskModal({
               onChange={(event) => setIsActive(event.target.checked)}
               className="edit-task-modal__status-checkbox h-4 w-4 rounded"
             />
-            <span className="edit-task-modal__status-label text-sm">{isActive ? t('editor.task.status.active') : t('editor.task.status.inactive')}</span>
+            <span className="edit-task-modal__status-label text-sm">
+              {isActive
+                ? t("editor.task.status.active")
+                : t("editor.task.status.inactive")}
+            </span>
           </label>
         </div>
       </section>
@@ -2122,7 +3017,9 @@ function EditTaskModal({
       {errors.user && <p className="text-xs text-rose-300">{errors.user}</p>}
       {errors.task && <p className="text-xs text-rose-300">{errors.task}</p>}
 
-      {toast && <ToastBanner tone={toast.type} message={toast.text} className="px-3" />}
+      {toast && (
+        <ToastBanner tone={toast.type} message={toast.text} className="px-3" />
+      )}
 
       <div className="flex flex-col gap-3 pt-2 sm:flex-row sm:justify-end">
         <button
@@ -2130,30 +3027,35 @@ function EditTaskModal({
           onClick={handleClose}
           className="edit-task-modal__button-secondary inline-flex items-center justify-center rounded-full border px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] transition"
         >
-          {t('editor.button.cancel')}
+          {t("editor.button.cancel")}
         </button>
         <button
           type="submit"
           disabled={isSubmitting}
           className="edit-task-modal__button-primary inline-flex items-center justify-center rounded-full px-5 py-2 text-sm font-semibold uppercase tracking-[0.18em] transition disabled:cursor-not-allowed disabled:opacity-60"
         >
-          {isSubmitting ? t('editor.button.saving') : t('editor.button.saveChanges')}
+          {isSubmitting
+            ? t("editor.button.saving")
+            : t("editor.button.saveChanges")}
         </button>
       </div>
     </div>
   );
 
-  if (variant === 'panel') {
+  if (variant === "panel") {
     return (
       <div className="fixed inset-0 z-[60] flex">
         <button
           type="button"
-          aria-label={t('editor.button.close')}
+          aria-label={t("editor.button.close")}
           onClick={handleClose}
           className="flex-1 bg-slate-950/60 backdrop-blur-sm"
         />
         <aside className="flex h-full w-full max-w-xl flex-col border-l border-[color:var(--color-border-subtle)] bg-slate-950/95 text-[color:var(--color-slate-100)] shadow-[0_18px_40px_rgba(15,23,42,0.65)]">
-          <form onSubmit={handleSubmit} className="flex h-full flex-col overflow-hidden">
+          <form
+            onSubmit={handleSubmit}
+            className="flex h-full flex-col overflow-hidden"
+          >
             <div className="flex items-center justify-between gap-3 border-b border-[color:var(--color-border-subtle)] px-6 py-4">
               {showNavigation ? (
                 <div className="flex items-center gap-2">
@@ -2163,7 +3065,7 @@ function EditTaskModal({
                     disabled={!previousTask}
                     className="rounded-full border border-white/15 px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:text-white disabled:cursor-not-allowed disabled:opacity-40"
                   >
-                    {t('editor.button.previous')}
+                    {t("editor.button.previous")}
                   </button>
                   <span className="text-[11px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-400)]">
                     {navigationLabel}
@@ -2174,18 +3076,20 @@ function EditTaskModal({
                     disabled={!nextTask}
                     className="rounded-full border border-white/15 px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:text-white disabled:cursor-not-allowed disabled:opacity-40"
                   >
-                    {t('editor.button.next')}
+                    {t("editor.button.next")}
                   </button>
                 </div>
               ) : (
-                <p className="text-sm font-semibold text-[color:var(--color-slate-200)]">{t('editor.modal.edit.panelTitle')}</p>
+                <p className="text-sm font-semibold text-[color:var(--color-slate-200)]">
+                  {t("editor.modal.edit.panelTitle")}
+                </p>
               )}
               <button
                 type="button"
                 onClick={handleClose}
                 className="rounded-full border border-white/15 px-3 py-1 text-[11px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:text-white"
               >
-                {t('editor.button.close')}
+                {t("editor.button.close")}
               </button>
             </div>
             <div className="flex-1 overflow-y-auto px-6 py-6">{formBody}</div>
@@ -2199,7 +3103,7 @@ function EditTaskModal({
     <div className="fixed inset-0 z-[60] flex items-end justify-center bg-slate-950/70 backdrop-blur-sm md:items-center">
       <button
         type="button"
-        aria-label={t('editor.button.close')}
+        aria-label={t("editor.button.close")}
         onClick={handleClose}
         className="absolute inset-0 h-full w-full"
       />
@@ -2227,6 +3131,214 @@ function ReadOnlyField({ label, value }: { label: string; value: string }) {
         {value}
       </div>
     </div>
+  );
+}
+
+function SuggestionsLabModal({
+  isOpen,
+  isApplying,
+  onClose,
+  onApply,
+}: {
+  isOpen: boolean;
+  isApplying: boolean;
+  onClose: () => void;
+  onApply: (taskIds: string[]) => void;
+}) {
+  const [selectedIds, setSelectedIds] = useState<string[]>([]);
+  const grouped = useMemo(() => {
+    return {
+      Body: EDITOR_LAB_QUICK_START_SEED.filter(
+        (task) => task.pillar === "Body",
+      ),
+      Mind: EDITOR_LAB_QUICK_START_SEED.filter(
+        (task) => task.pillar === "Mind",
+      ),
+      Soul: EDITOR_LAB_QUICK_START_SEED.filter(
+        (task) => task.pillar === "Soul",
+      ),
+    };
+  }, []);
+  const totalAvailableSuggestions = EDITOR_LAB_QUICK_START_SEED.length;
+  const selectedSet = useMemo(() => new Set(selectedIds), [selectedIds]);
+  const selectedCount = selectedIds.length;
+  const hasSelectedSuggestions = selectedCount > 0;
+  const isEverythingSelected = selectedCount === totalAvailableSuggestions;
+
+  useEffect(() => {
+    if (!isOpen) {
+      setSelectedIds([]);
+    }
+  }, [isOpen]);
+
+  if (!isOpen) {
+    return null;
+  }
+
+  return (
+    <div className="editor-suggestions-overlay fixed inset-0 z-[75] flex items-end justify-center bg-slate-950/80 backdrop-blur-md md:items-center">
+      <button
+        type="button"
+        aria-label="Cerrar sugerencias"
+        onClick={onClose}
+        className="absolute inset-0 h-full w-full"
+      />
+      <section className="editor-suggestions-modal relative z-10 w-full max-w-3xl rounded-t-3xl border border-[color:var(--color-border-subtle)] bg-[linear-gradient(180deg,rgba(15,23,42,0.95),rgba(15,23,42,0.92))] p-4 shadow-[0_28px_80px_rgba(2,6,23,0.5)] md:rounded-3xl md:p-6">
+        <div className="flex items-start justify-between gap-4">
+          <div>
+            <p className="text-[11px] font-semibold uppercase tracking-[0.24em] text-violet-200/85">
+              Sugerencias
+            </p>
+            <h2 className="mt-2 text-lg font-semibold text-white md:text-xl">
+              Activá tareas recomendadas
+            </h2>
+            <p className="mt-2 max-w-xl text-sm leading-relaxed text-[color:var(--color-slate-300)]">
+              Seleccioná varias sugerencias por pilar y agregalas en lote a tu sistema.
+            </p>
+          </div>
+          <button
+            type="button"
+            onClick={onClose}
+            aria-label="Cerrar sugerencias"
+            className="inline-flex h-9 w-9 items-center justify-center rounded-full border border-white/15 bg-white/5 text-[color:var(--color-slate-200)] transition hover:border-white/35 hover:text-white"
+          >
+            <CloseTinyIcon className="h-4 w-4" />
+          </button>
+        </div>
+        <div className="mt-4 flex flex-wrap items-center justify-between gap-2 rounded-2xl border border-white/10 bg-white/[0.02] px-3 py-2">
+          <p className="text-xs text-[color:var(--color-slate-300)]">
+            {selectedCount} de {totalAvailableSuggestions} seleccionadas
+          </p>
+          <div className="flex items-center gap-1.5">
+            <button
+              type="button"
+              onClick={() =>
+                setSelectedIds(
+                  isEverythingSelected
+                    ? []
+                    : EDITOR_LAB_QUICK_START_SEED.map((task) => task.id),
+                )
+              }
+              className="rounded-full border border-white/15 px-3 py-1 text-[10px] font-semibold uppercase tracking-[0.2em] text-[color:var(--color-slate-200)] transition hover:border-white/30 hover:text-white"
+            >
+              {isEverythingSelected ? "Limpiar" : "Seleccionar todo"}
+            </button>
+          </div>
+        </div>
+        <div className="mt-4 max-h-[58vh] space-y-5 overflow-y-auto pr-1 pb-24 md:pb-20">
+          {(
+            Object.entries(grouped) as Array<
+              [keyof typeof grouped, typeof EDITOR_LAB_QUICK_START_SEED]
+            >
+          ).map(([pillar, tasks]) => (
+            <div key={pillar} className="space-y-2.5">
+              <div className="flex items-center justify-between gap-3">
+                <h3 className="text-xs font-semibold uppercase tracking-[0.22em] text-[color:var(--color-slate-400)]">
+                  {pillar}
+                </h3>
+                <p className="text-[10px] uppercase tracking-[0.2em] text-[color:var(--color-slate-500)]">
+                  {
+                    tasks.filter((task) => selectedSet.has(task.id)).length
+                  }{" "}
+                  / {tasks.length}
+                </p>
+              </div>
+              <div className="grid gap-2 sm:grid-cols-2">
+                {tasks.map((task) => {
+                  const checked = selectedSet.has(task.id);
+                  return (
+                    <button
+                      key={task.id}
+                      type="button"
+                      onClick={() =>
+                        setSelectedIds((current) =>
+                          current.includes(task.id)
+                            ? current.filter((id) => id !== task.id)
+                            : [...current, task.id],
+                        )
+                      }
+                      aria-pressed={checked}
+                      className={`group flex items-start gap-3 rounded-2xl border px-3 py-3 text-left transition ${checked ? "border-violet-300/60 bg-[linear-gradient(155deg,rgba(167,139,250,0.25),rgba(129,140,248,0.12))] shadow-[0_10px_24px_rgba(139,92,246,0.24)]" : "border-[color:var(--color-border-subtle)] bg-[color:var(--color-overlay-1)] hover:border-white/25"}`}
+                    >
+                      <span
+                        className={`mt-0.5 inline-flex h-5 w-5 shrink-0 items-center justify-center rounded-full border text-[10px] transition ${checked ? "border-violet-100 bg-violet-200/90 text-violet-700" : "border-white/25 text-transparent group-hover:border-white/40"}`}
+                      >
+                        ✓
+                      </span>
+                      <span className="flex-1">
+                        <span className="block text-sm font-medium text-[color:var(--color-slate-100)]">
+                          {task.title}
+                        </span>
+                        <span className="mt-1 inline-flex rounded-full border border-violet-200/20 bg-violet-400/10 px-2 py-0.5 text-[9px] font-medium uppercase tracking-[0.14em] text-violet-100/85">
+                          {task.trait}
+                        </span>
+                      </span>
+                    </button>
+                  );
+                })}
+              </div>
+            </div>
+          ))}
+        </div>
+        <div className="absolute inset-x-0 bottom-0 flex items-center justify-between gap-3 border-t border-white/10 bg-[linear-gradient(180deg,rgba(15,23,42,0.05),rgba(15,23,42,0.96)_24%)] px-4 py-4 backdrop-blur-md md:px-6">
+          <p className="text-xs text-[color:var(--color-slate-400)]">
+            {hasSelectedSuggestions
+              ? `Listo para sumar ${selectedCount} tarea(s) a tu sistema.`
+              : "Seleccioná tareas para empezar a construir tu sistema."}
+          </p>
+          <button
+            type="button"
+            disabled={!hasSelectedSuggestions || isApplying}
+            onClick={() => onApply(selectedIds)}
+            className="inline-flex rounded-full bg-violet-500 px-5 py-2.5 text-xs font-semibold uppercase tracking-[0.18em] text-white shadow-[0_12px_30px_rgba(139,92,246,0.35)] transition hover:bg-violet-400 disabled:cursor-not-allowed disabled:opacity-50"
+          >
+            {isApplying
+              ? "Sumando tareas…"
+              : hasSelectedSuggestions
+                ? `Sumar ${selectedCount} a mi sistema`
+                : "Sumar a mi sistema"}
+          </button>
+        </div>
+      </section>
+    </div>
+  );
+}
+
+function GuideCompassIcon({ className }: { className?: string }) {
+  return (
+    <svg viewBox="0 0 24 24" fill="none" aria-hidden="true" className={className}>
+      <circle cx="12" cy="12" r="8.5" stroke="currentColor" strokeWidth="1.8" opacity="0.75" />
+      <path
+        d="M9.2 14.8 10.8 9.2 14.8 10.8 13.2 14.8 9.2 14.8Z"
+        stroke="currentColor"
+        strokeWidth="1.8"
+        strokeLinejoin="round"
+      />
+      <path d="M10.8 9.2 14.8 10.8" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" />
+    </svg>
+  );
+}
+
+function SuggestionsMagicIcon({ className }: { className?: string }) {
+  return (
+    <svg viewBox="0 0 24 24" fill="none" aria-hidden="true" className={className}>
+      <path
+        d="M12 4.5a5.5 5.5 0 0 0-3.3 9.9c.5.37.8.94.8 1.56V17h5v-1.04c0-.62.3-1.19.8-1.56A5.5 5.5 0 0 0 12 4.5Z"
+        stroke="currentColor"
+        strokeWidth="1.6"
+        strokeLinejoin="round"
+      />
+      <path d="M9.7 17h4.6v1.45a1.15 1.15 0 0 1-1.15 1.15h-2.3a1.15 1.15 0 0 1-1.15-1.15V17Z" fill="currentColor" />
+      <path d="M10.1 21h3.8" stroke="currentColor" strokeWidth="1.6" strokeLinecap="round" />
+    </svg>
+  );
+}
+
+function CloseTinyIcon({ className }: { className?: string }) {
+  return (
+    <svg viewBox="0 0 24 24" fill="none" aria-hidden="true" className={className}>
+      <path d="m7 7 10 10M17 7 7 17" stroke="currentColor" strokeWidth="2" strokeLinecap="round" />
+    </svg>
   );
 }
